refactor(output): split html.py monolith into html/ package
Closes #1522 (final phase). The 3306-LOC output/html.py is broken into
cohesive submodules with zero behaviour change; the package re-exports every
public and semi-public symbol so importers are unaffected:

  html/__init__.py   re-exports (to_html, export_html, _vuln_table,
                     _cis_benchmark_section, _PAGE_SIZE, ...)
  html/_common.py    colour constants + _esc / _sev_badge
  html/sections.py   section renderers (summary/tables/graph-data/CIS/compliance)
  html/styles.py     embedded CSS/theme
  html/scripts.py    embedded JS (tabs/pagination/graph) + offline post-proc
  html/tabs.py       tab-bar post-processing
  html/document.py   top-level assembler + file exporter

Byte-for-byte identical HTML verified for a fixture report in both online and
offline modes (diff empty). ruff + mypy clean; per-file E501/F541 ignore moved
to the new package path (the f-prefix on the static script string is
load-bearing — it unescapes doubled braces).
