--- src/agent_bom/output/html.py
@@ -1,3306 +1,0 @@
-"""HTML report generator for AI-BOM scans.
-
-Produces a single ``report.html`` file with:
-- Enterprise dashboard: stat cards, severity donut, blast radius bar chart
-- Hierarchical Cytoscape.js risk map (dagre layout: Provider → Agent → Server → Package → CVE)
-  with zoom controls, fullscreen toggle, and node highlighting
-- Collapsible agent inventory panels with search and truncated package lists
-- Sortable vulnerability table with severity pill, CVSS bar, EPSS, KEV badge
-- Blast radius table ordered by risk score with visual bar
-- Remediation plan ordered by impact
-- Skill audit findings section (when skill scan data is present)
-- Print-friendly stylesheet for PDF export
-
-No server required — open the file in any browser. By default, Chart.js,
-Cytoscape.js, and dagre load from pinned public CDNs. Pass
-``offline_assets=True`` to emit a static airgap-safe report with no external
-script dependencies.
-"""
-
-from __future__ import annotations
-
-import json
-from pathlib import Path
-from typing import TYPE_CHECKING
-
-from agent_bom.finding import FindingType
-from agent_bom.graph.severity import SEVERITY_THRESHOLD_LABELS, severity_policy_rank, severity_worst_first_rank
-from agent_bom.output.exposure_path import exposure_path_brief_for_finding
-from agent_bom.output.finding_views import (
-    compliance_row_dict,
-    cve_findings,
-    evidence,
-    exploit_likelihood_value,
-    package_name,
-    package_version,
-    ranked_cve_findings,
-    severity_value,
-    topology_package_key,
-)
-from agent_bom.security import sanitize_launch_command, sanitize_path_label
-
-if TYPE_CHECKING:
-    from agent_bom.finding import Finding
-    from agent_bom.models import AIBOMReport, BlastRadius
-
-
-_SEV_COLOR = {
-    "critical": "#dc2626",
-    "high": "#ea580c",
-    "medium": "#d97706",
-    "low": "#6b7280",
-    "none": "#16a34a",
-    "unknown": "#9ca3af",
-}
-
-# Max packages shown per server before collapsing
-_PKG_PREVIEW = 15
-
-# Client-side pagination: rows shown per page for large findings tables.
-_PAGE_SIZE = 50
-
-# Tab grouping for the report. Each entry: (tab-key, label, [section ids]).
-# Only tabs whose sections are actually rendered appear in the tab bar, so a
-# small scan collapses to a couple of tabs while a large multi-cloud scan gets
-# the full set. Order here is the tab order left-to-right.
-_TAB_DEFS: tuple[tuple[str, str, tuple[str, ...]], ...] = (
-    ("summary", "Summary", ("summary", "charts")),
-    ("agents", "Agents &amp; Servers", ("riskmap", "inventory", "aiinventory")),
-    (
-        "findings",
-        "CVEs &amp; Findings",
-        ("attackflow", "vulns", "policyfindings", "exposure-paths", "blast", "remediation"),
-    ),
-    ("compliance", "CIS &amp; Compliance", ("compliance", "cisbenchmarks", "skillaudit")),
-    ("governance", "Trust &amp; Governance", ("trust", "enforcement")),
-)
-
-
-def _pager_controls(table_id: str, total: int, page_size: int = _PAGE_SIZE) -> str:
-    """Server-rendered pagination bar bound to a table by ``data-pager``.
-
-    Rendered for every findings table so the controls exist in the static file
-    (and in tests); the paginator JS wires the buttons and hides the whole bar
-    when the filtered set fits on a single page.
-    """
-    return (
-        f'<div class="pager" data-pager="{table_id}" data-page-size="{page_size}">'
-        '<button class="pager-btn" data-act="first" title="First page">&laquo;</button>'
-        '<button class="pager-btn" data-act="prev" title="Previous page">&lsaquo; Prev</button>'
-        f'<span class="pager-info">1&ndash;{min(page_size, total)} of {total}</span>'
-        '<button class="pager-btn" data-act="next" title="Next page">Next &rsaquo;</button>'
-        '<button class="pager-btn" data-act="last" title="Last page">&raquo;</button>'
-        '<span class="pager-sep"></span>'
-        '<label class="pager-size-label">Rows'
-        '<select class="pager-size">'
-        '<option value="25">25</option>'
-        f'<option value="50"{" selected" if page_size == 50 else ""}>50</option>'
-        '<option value="100">100</option>'
-        '<option value="250">250</option>'
-        '</select></label>'
-        "</div>"
-    )
-
-
-def _apply_tabs(html: str, tab_counts: dict[str, int]) -> str:
-    """Reorganise the flat report into JS-driven tabs.
-
-    Tags each ``<section id=...>`` with its ``data-tab`` group, injects a tab
-    bar at the top of the container, and relies on CSS (``body.js-tabs``) plus a
-    small script (added in the main IIFE) to show one tab at a time. Degrades to
-    the full scrollable page when JS is off or when printing.
-    """
-    present: list[tuple[str, str]] = []
-    for tab_key, label, section_ids in _TAB_DEFS:
-        has_any = False
-        for sid in section_ids:
-            needle = f'<section id="{sid}"'
-            if needle in html:
-                html = html.replace(needle, f'<section id="{sid}" data-tab="{tab_key}"', 1)
-                has_any = True
-        if has_any:
-            present.append((tab_key, label))
-
-    if len(present) < 2:  # nothing worth tabbing
-        return html
-
-    buttons = []
-    for idx, (tab_key, label) in enumerate(present):
-        count = tab_counts.get(tab_key, 0)
-        badge = f'<span class="tab-count">{count}</span>' if count else ""
-        active = " active" if idx == 0 else ""
-        buttons.append(
-            f'<button class="tab-btn{active}" data-tab="{tab_key}" '
-            f'role="tab">{label}{badge}</button>'
-        )
-    tab_bar = '<nav class="tab-bar" role="tablist">' + "".join(buttons) + "</nav>\n"
-
-    marker = '<div class="container">\n'
-    return html.replace(marker, marker + tab_bar, 1)
-
-_EXTERNAL_SCRIPT_TAGS = (
-    '  <script src="https://cdn.jsdelivr.net/npm/chart.js@4.4.2/dist/chart.umd.min.js"></script>\n',
-    '  <script src="https://unpkg.com/cytoscape@3.30.2/dist/cytoscape.min.js"></script>\n',
-    '  <script src="https://unpkg.com/dagre@0.8.5/dist/dagre.min.js"></script>\n',
-    '  <script src="https://unpkg.com/cytoscape-dagre@2.5.0/cytoscape-dagre.js"></script>\n',
-    '  <script src="https://unpkg.com/cytoscape-popper@2.0.0/cytoscape-popper.js"></script>\n',
-)
-
-
-def _sev_badge(sev: str) -> str:
-    color = _SEV_COLOR.get(sev.lower(), "#6b7280")
-    return (
-        f'<span style="background:{color};color:#fff;padding:2px 8px;border-radius:4px;'
-        f'font-size:.72rem;font-weight:700;letter-spacing:.04em">{sev.upper()}</span>'
-    )
-
-
-def _esc(s: object) -> str:
-    return str(s).replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;").replace('"', "&quot;")
-
-
-# ─── Data builders ────────────────────────────────────────────────────────────
-
-
-def _chart_data(findings: list["Finding"]) -> str:
-    """Build Chart.js dataset JSON for severity donut + blast radius bar chart."""
-    from agent_bom.models import Severity
-
-    sev_counts: dict[str, int] = {s.value: 0 for s in Severity}
-    for finding in findings:
-        sev = severity_value(finding)
-        if sev in sev_counts:
-            sev_counts[sev] += 1
-
-    top10 = sorted(findings, key=lambda finding: float(finding.risk_score or 0.0), reverse=True)[:10]
-    blast_labels = [
-        f"{(finding.cve_id or finding.title)[:16]}/{package_name(finding)[:14]}"
-        for finding in top10
-    ]
-    blast_scores = [round(float(finding.risk_score or 0.0), 2) for finding in top10]
-    blast_colors = [_SEV_COLOR.get(severity_value(finding), "#6b7280") for finding in top10]
-
-    return json.dumps(
-        {
-            "sev": {
-                "labels": [k.capitalize() for k in sev_counts],
-                "data": list(sev_counts.values()),
-                "colors": [_SEV_COLOR[k] for k in sev_counts],
-            },
-            "blast": {
-                "labels": blast_labels,
-                "scores": blast_scores,
-                "colors": blast_colors,
-            },
-        }
-    )
-
-
-def _cytoscape_elements(report: "AIBOMReport", blast_radii: list["BlastRadius"]) -> str:
-    """Build Cytoscape element list using the shared graph builder."""
-    from agent_bom.output.graph import build_graph_elements
-
-    elements = build_graph_elements(report, blast_radii, include_cve_nodes=True)
-    return json.dumps(elements)
-
-
-def _attack_flow_elements(report: "AIBOMReport", blast_radii: list["BlastRadius"]) -> str:
-    """Build attack flow element list showing CVE → impact propagation."""
-    from agent_bom.output.graph import build_attack_flow_elements
-
-    elements = build_attack_flow_elements(report, blast_radii)
-    return json.dumps(elements)
-
-
-# ─── HTML sections ────────────────────────────────────────────────────────────
-
-
-def _delta_banner(report: "AIBOMReport") -> str:
-    """Render a delta mode info banner when the scan used --delta."""
-    scan_json = getattr(report, "_cached_json", None)
-    delta = (scan_json or {}).get("delta") if scan_json else None
-    if not delta or not delta.get("enabled"):
-        return ""
-    new_count = delta.get("new_count", 0)
-    pre_count = delta.get("pre_existing_count", 0)
-    baseline = _esc(str(delta.get("baseline_path") or ""))
-    parts = []
-    if new_count:
-        parts.append(f"<b>{new_count}</b> new finding(s)")
-    if pre_count:
-        parts.append(f"<b>{pre_count}</b> pre-existing (suppressed from exit code)")
-    body = " &middot; ".join(parts) if parts else "No new findings"
-    baseline_html = f' &nbsp;<span style="color:#94a3b8;font-size:.75rem">baseline: {baseline}</span>' if baseline else ""
-    return (
-        '<div style="background:#1e3a5f;border:1px solid #2563eb;border-radius:8px;'
-        'padding:10px 16px;margin-bottom:12px;font-size:.85rem;color:#93c5fd">'
-        f"&#x1f504; <b>Delta mode</b>: {body}{baseline_html}"
-        "</div>"
-    )
-
-
-def _warn_gate_banner(report: "AIBOMReport") -> str:
-    """Render a warn-gate banner when findings triggered the warn threshold."""
-    scan_json = getattr(report, "_cached_json", None)
-    if not scan_json:
-        return ""
-    if scan_json.get("warn_gate_status") != "warn":
-        return ""
-    count = scan_json.get("warn_gate_count", 0)
-    sev = _esc(str(scan_json.get("warn_gate_severity") or ""))
-    return (
-        '<div style="background:#422006;border:1px solid #d97706;border-radius:8px;'
-        'padding:10px 16px;margin-bottom:12px;font-size:.85rem;color:#fcd34d">'
-        f"&#x26a0;&#xfe0f; <b>Warn gate triggered</b>: {count} finding(s) at or above "
-        f"<b>{sev}</b> severity &mdash; exit 0 (warning only, not a hard failure)"
-        "</div>"
-    )
-
-
-def _summary_cards(report: "AIBOMReport", findings: list["Finding"], policy_findings: list["Finding"]) -> str:
-    crit = sum(1 for finding in findings if severity_value(finding) == "critical")
-    policy_crit = sum(1 for finding in policy_findings if str(finding.severity).lower() == "critical")
-    total_vulns = len(findings)
-    cred_servers = sum(1 for a in report.agents for s in a.mcp_servers if s.has_credentials)
-    kev_count = sum(1 for finding in findings if finding.is_kev)
-
-    def card(icon: str, value: str, label: str, accent: str, sub: str = "") -> str:
-        sub_html = f'<div style="font-size:.68rem;color:#475569;margin-top:2px">{sub}</div>' if sub else ""
-        return (
-            f'<div class="stat-card" style="border-left-color:{accent}">'
-            f'<div class="stat-icon">{icon}</div>'
-            f'<div class="stat-value" style="color:{accent}">{_esc(value)}</div>'
-            f'<div class="stat-label">{label}</div>'
-            f"{sub_html}"
-            f"</div>"
-        )
-
-    return (
-        '<div class="stat-grid">'
-        + "".join(
-            [
-                card("&#x1f916;", str(report.total_agents), "Agents", "#60a5fa", f"{report.total_servers} servers"),
-                card("&#x1f4e6;", str(report.total_packages), "Packages", "#38bdf8", "direct + transitive"),
-                card("&#x26a0;&#xfe0f;", str(total_vulns), "Vulnerabilities", "#f87171" if total_vulns else "#34d399", "across all agents"),
-                card(
-                    "&#x1f6e1;&#xfe0f;",
-                    str(len(policy_findings)),
-                    "Policy Findings",
-                    "#f87171" if policy_findings else "#34d399",
-                    "non-CVE controls",
-                ),
-                card("&#x1f511;", str(cred_servers), "Servers w/ Creds", "#fbbf24" if cred_servers else "#34d399", "credential exposure"),
-                card(
-                    "&#x1f6a8;", str(crit + policy_crit), "Critical", "#ef4444" if crit or policy_crit else "#34d399", "needs immediate fix"
-                ),
-                card("&#x1f9a0;", str(kev_count), "CISA KEV", "#a855f7" if kev_count else "#34d399", "actively exploited"),
-            ]
-        )
-        + "</div>"
-    )
-
-
-def _vuln_table(report: "AIBOMReport", blast_radii: list["BlastRadius"]) -> str:
-    findings = cve_findings(report, blast_radii)
-    if not findings:
-        return '<div class="empty-state">&#x2705; No vulnerabilities found in scanned packages.</div>'
-
-    has_missing = any(not finding.cvss_score or not finding.description for finding in findings)
-    hint = ""
-    if has_missing:
-        hint = (
-            '<div class="hint-box">'
-            "&#x1f4a1; <strong>Some entries are missing CVSS scores or descriptions.</strong> "
-            "Run with <code>--enrich</code> to fetch full NVD metadata, CVSS 3.x vectors, EPSS, and CISA KEV status."
-            "</div>"
-        )
-
-    sorted_findings = sorted(
-        findings,
-        key=lambda finding: severity_policy_rank(severity_value(finding)),
-        reverse=True,
-    )
-    rows = []
-    for idx, finding in enumerate(sorted_findings):
-        sev = severity_value(finding)
-        color = _SEV_COLOR.get(sev, "#6b7280")
-        cvss_bar = ""
-        if finding.cvss_score:
-            pct = int(finding.cvss_score * 10)
-            cvss_bar = (
-                f'<div style="display:flex;align-items:center;gap:6px">'
-                f'<div style="background:#0f172a;border-radius:3px;height:4px;width:52px;flex-shrink:0">'
-                f'<div style="background:{color};border-radius:3px;height:4px;width:{pct}%"></div></div>'
-                f'<strong style="color:{color}">{finding.cvss_score:.1f}</strong></div>'
-            )
-        else:
-            cvss_bar = '<span style="color:#334155">&mdash;</span>'
-        epss = f"{finding.epss_score:.1%}" if finding.epss_score else '<span style="color:#334155">&mdash;</span>'
-        exploit_level = exploit_likelihood_value(finding)
-        if finding.is_kev:
-            kev = '<span class="badge-kev" title="CISA Known Exploited Vulnerability">KEV</span>'
-        elif exploit_level == "likely_exploited":
-            kev = '<span class="badge-exploit-likely" title="EPSS ≥ 0.5 or percentile ≥ 95 — exploitation likely">EXPL</span>'
-        elif exploit_level == "public_exploit":
-            kev = '<span class="badge-exploit-public" title="EPSS percentile ≥ 80 — public exploit code">PoC</span>'
-        else:
-            kev = '<span style="color:#334155">&mdash;</span>'
-        fix = (
-            f'<code style="color:#4ade80">{_esc(finding.fixed_version)}</code>'
-            if finding.fixed_version
-            else '<span style="color:#475569">No fix</span>'
-        )
-        summary_text = (finding.description or "")[:90]
-        summary = _esc(summary_text) if summary_text else '<span style="color:#475569;font-style:italic">Run --enrich</span>'
-        agents_s = ", ".join(_esc(name) for name in finding.affected_agents) or "<span style='color:#334155'>&mdash;</span>"
-        creds_s = (
-            " ".join(f'<code style="color:#fbbf24">{_esc(c)}</code>' for c in finding.exposed_credentials)
-            or "<span style='color:#334155'>&mdash;</span>"
-        )
-        vendor_sev = finding.vendor_severity
-        vendor_hint = ""
-        if vendor_sev and vendor_sev.lower() != sev:
-            vendor_hint = f'<br><span style="font-size:.62rem;color:#94a3b8;font-style:italic">vendor: {_esc(vendor_sev)}</span>'
-        tier = evidence(finding, "match_confidence_tier", None)
-        tier_hint = ""
-        if tier:
-            _tier_color = "#f59e0b" if tier == "nvd_cpe_candidate" else "#64748b"
-            tier_hint = (
-                f'<br><span class="match-tier" data-tier="{_esc(tier)}" '
-                f'title="match confidence: {_esc(tier)}" '
-                f'style="font-size:.6rem;color:{_tier_color}">{_esc(str(tier).replace("_", " "))}</span>'
-            )
-        vuln_id = finding.cve_id or finding.id
-        pkg_name = package_name(finding)
-        pkg_version = package_version(finding)
-        pg_cls = " pg-hidden" if idx >= _PAGE_SIZE else ""
-        rows.append(
-            f'<tr class="pg-row{pg_cls}" data-severity="{sev}" data-kev="{"1" if finding.is_kev else "0"}" '
-            f'data-exploit-likelihood="{exploit_level}" '
-            f'data-match-tier="{_esc(tier or "")}" '
-            f'data-cvss="{finding.cvss_score if finding.cvss_score else 0}">'
-            f'<td><code class="vuln-id">{_esc(vuln_id)}</code>{tier_hint}</td>'
-            f"<td>{_sev_badge(sev)}{vendor_hint}</td>"
-            f'<td><strong style="color:#e2e8f0">{_esc(pkg_name)}</strong>'
-            f'<span style="color:#475569;font-size:.78rem">@{_esc(pkg_version)}</span></td>'
-            f"<td>{cvss_bar}</td>"
-            f'<td style="text-align:center;font-size:.82rem;color:#94a3b8">{epss}</td>'
-            f'<td style="text-align:center">{kev}</td>'
-            f"<td>{fix}</td>"
-            f'<td style="font-size:.78rem;color:#94a3b8">{agents_s}</td>'
-            f'<td style="font-size:.78rem">{creds_s}</td>'
-            f'<td style="font-size:.75rem;color:#64748b;max-width:180px">{summary}</td>'
-            f"</tr>"
-        )
-
-    headers = ["Vuln ID", "Severity", "Package", "CVSS", "EPSS", "KEV", "Fix", "Affected Agents", "Exposed Creds", "Summary"]
-
-    filter_bar = (
-        '<div class="vuln-filter-bar" style="display:flex;flex-wrap:wrap;gap:12px;align-items:center;'
-        'margin-bottom:14px;padding:12px 16px;background:#0f172a;border-radius:8px;border:1px solid #1e293b">'
-        '<span style="font-size:.72rem;color:#64748b;text-transform:uppercase;letter-spacing:.06em;font-weight:700">Filter:</span>'
-        '<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:#fca5a5;cursor:pointer">'
-        '<input type="checkbox" class="vuln-sev-filter" value="critical" checked> Critical</label>'
-        '<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:#fb923c;cursor:pointer">'
-        '<input type="checkbox" class="vuln-sev-filter" value="high" checked> High</label>'
-        '<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:#fbbf24;cursor:pointer">'
-        '<input type="checkbox" class="vuln-sev-filter" value="medium" checked> Medium</label>'
-        '<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:#94a3b8;cursor:pointer">'
-        '<input type="checkbox" class="vuln-sev-filter" value="low" checked> Low</label>'
-        '<span style="width:1px;height:18px;background:#334155"></span>'
-        '<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:#fca5a5;cursor:pointer">'
-        '<input type="checkbox" id="kevToggle"> KEV only</label>'
-        '<span style="width:1px;height:18px;background:#334155"></span>'
-        '<input type="text" id="vulnSearch" placeholder="Search vulns&hellip;" '
-        'style="padding:6px 10px;background:#1e293b;border:1px solid #334155;border-radius:6px;'
-        'color:#e2e8f0;font-size:.78rem;width:160px;outline:none">'
-        "</div>"
-    )
-
-    return (
-        hint
-        + filter_bar
-        + '<div class="table-wrap"><table class="data-table sortable paginated" id="vulnTable">'
-        + "<thead><tr>"
-        + "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-        + "</tr></thead>"
-        + f"<tbody>{''.join(rows)}</tbody></table></div>"
-        + _pager_controls("vulnTable", len(rows))
-    )
-
-
-def _blast_table(report: "AIBOMReport", blast_radii: list["BlastRadius"]) -> str:
-    findings = cve_findings(report, blast_radii)
-    if not findings:
-        return ""
-    sorted_findings = sorted(findings, key=lambda finding: float(finding.risk_score or 0.0), reverse=True)
-    rows = []
-    for i, finding in enumerate(sorted_findings, 1):
-        sev = severity_value(finding)
-        color = _SEV_COLOR.get(sev, "#6b7280")
-        risk_score = float(finding.risk_score or 0.0)
-        bar_w = int(risk_score * 9)
-        ai_badge = '<span class="badge-ai">AI</span>' if finding.ai_risk_context else ""
-        kev_badge = '<span class="badge-kev">KEV</span>' if finding.is_kev else ""
-        fix = (
-            f'<code style="color:#4ade80;font-size:.8rem">{_esc(finding.fixed_version)}</code>'
-            if finding.fixed_version
-            else '<span style="color:#475569">&mdash;</span>'
-        )
-        vuln_id = finding.cve_id or finding.id
-        rows.append(
-            f"<tr>"
-            f'<td style="color:#475569;font-weight:600">#{i}</td>'
-            f'<td><code class="vuln-id">{_esc(vuln_id)}</code></td>'
-            f"<td>{_sev_badge(sev)}</td>"
-            f"<td>"
-            f'<div style="display:flex;align-items:center;gap:8px">'
-            f'<div style="background:#0f172a;border-radius:3px;height:5px;width:90px">'
-            f'<div style="background:{color};border-radius:3px;height:5px;width:{bar_w}px"></div></div>'
-            f'<strong style="color:{color}">{risk_score:.1f}</strong></div>'
-            f"</td>"
-            f'<td style="text-align:center;color:#e2e8f0">{len(finding.affected_agents)}</td>'
-            f'<td style="text-align:center;color:#fbbf24">{len(finding.exposed_credentials)}</td>'
-            f'<td style="text-align:center;color:#94a3b8">{len(finding.exposed_tools)}</td>'
-            f"<td>{ai_badge}{kev_badge}</td>"
-            f"<td>{fix}</td>"
-            f"</tr>"
-        )
-    headers = ["#", "Vuln ID", "Severity", "Blast Score (0&ndash;10)", "Agents Hit", "Creds Exposed", "Tools Reachable", "Flags", "Fix"]
-    return (
-        '<div class="table-wrap"><table class="data-table sortable">'
-        + "<thead><tr>"
-        + "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-        + f"</tr></thead><tbody>{''.join(rows)}</tbody></table></div>"
-    )
-
-
-def _exposure_path_section(report: "AIBOMReport", blast_radii: list["BlastRadius"]) -> str:
-    findings = ranked_cve_findings(report, blast_radii)
-    if not findings:
-        return ""
-    cards = []
-    for rank, finding in enumerate(findings, 1):
-        brief = exposure_path_brief_for_finding(finding, rank=rank)
-        sev = str(finding.effective_severity() or finding.severity or "unknown").lower()
-        color = _SEV_COLOR.get(sev, "#6b7280")
-        cards.append(
-            '<div class="exposure-path-card">'
-            f'<div class="path-rank" style="border-color:{color};color:{color}">#{_esc(brief["rank"])}</div>'
-            '<div class="path-body">'
-            f'<div class="path-title">{_esc(brief["path"])}</div>'
-            f'<div class="path-summary">{_esc(brief["why"])}</div>'
-            '<div class="path-meta">'
-            f'<span style="color:{color};font-weight:700">{_esc(brief["severity"])}</span>'
-            f"<span>risk {_esc(brief['risk'])}</span>"
-            f"<span>{_esc(brief['proof'])}</span>"
-            f"</div>"
-            f'<div class="path-fix">{_esc(brief["fix"])}</div>'
-            "</div>"
-            "</div>"
-        )
-    return "".join(cards)
-
-
-def _remediation_list(findings: list["Finding"]) -> str:
-    if not findings:
-        return '<p style="color:#4ade80">&#x2705; Nothing to remediate.</p>'
-    with_fix = sorted(
-        [finding for finding in findings if finding.fixed_version],
-        key=lambda finding: float(finding.risk_score or 0.0),
-        reverse=True,
-    )
-    no_fix = [finding for finding in findings if not finding.fixed_version]
-    items = []
-    for finding in with_fix:
-        creds_note = (
-            f' &middot; frees <strong style="color:#fbbf24">{len(finding.exposed_credentials)}</strong> credential(s)'
-            if finding.exposed_credentials
-            else ""
-        )
-        items.append(
-            f'<div class="remediation-item">'
-            f'<div style="flex-shrink:0;padding-top:1px">{_sev_badge(severity_value(finding))}</div>'
-            f'<div style="flex:1">'
-            f'<div style="color:#e2e8f0;font-weight:600">{_esc(package_name(finding))}'
-            f'<span style="color:#475569;font-weight:400">@{_esc(package_version(finding))}</span></div>'
-            f'<div style="font-size:.8rem;color:#64748b;margin-top:3px">'
-            f'<code class="vuln-id">{_esc(finding.cve_id or finding.title)}</code>'
-            f' &middot; upgrade to <code style="color:#4ade80">{_esc(finding.fixed_version)}</code>'
-            f" &middot; protects <strong>{len(finding.affected_agents)}</strong> agent(s)"
-            f"{creds_note}"
-            f"</div></div>"
-            f'<div style="flex-shrink:0;color:#475569;font-size:.78rem;padding-top:3px">score&nbsp;{float(finding.risk_score or 0.0):.1f}</div>'
-            f"</div>"
-        )
-    nf_html = ""
-    if no_fix:
-        nf_rows = "".join(
-            f'<div style="padding:9px 0;border-bottom:1px solid #1e293b;font-size:.82rem">'
-            f"{_sev_badge(severity_value(finding))} "
-            f'<code class="vuln-id">{_esc(finding.cve_id or finding.title)}</code> &mdash; '
-            f'<strong style="color:#e2e8f0">{_esc(package_name(finding))}</strong>@{_esc(package_version(finding))}'
-            f' &mdash; <span style="color:#475569">no fix available &mdash; monitor upstream</span></div>'
-            for finding in no_fix
-        )
-        nf_html = '<div style="margin-top:20px"><div class="subsection-label">No Fix Available</div>' + nf_rows + "</div>"
-    return "".join(items) + nf_html
-
-
-def _non_cve_findings(report: "AIBOMReport") -> list["Finding"]:
-    """Return unified findings not already represented in the CVE or CIS tables.
-
-    Cloud CIS benchmark FAILures are lifted into the unified stream by
-    ``to_findings()`` AND rendered per-check by the dedicated CIS Benchmark
-    Posture table (``_CIS_CLOUD_LABELS``). Skip the unified copy for any
-    provider that table covers so each failed check renders exactly once (no
-    double-emit). Snowflake governance findings are also ``CIS_FAIL`` but carry
-    no ``benchmark`` marker and have no dedicated table, so they stay here.
-    """
-    findings: list[Finding] = []
-    for finding in report.to_findings():
-        if finding.finding_type == FindingType.CVE:
-            continue
-        evidence = finding.evidence if isinstance(finding.evidence, dict) else {}
-        if (
-            finding.finding_type == FindingType.CIS_FAIL
-            and evidence.get("benchmark") == "CIS"
-            and evidence.get("provider") in _CIS_CLOUD_LABELS
-        ):
-            continue
-        findings.append(finding)
-    return findings
-
-
-def _policy_findings_section(findings: list["Finding"]) -> str:
-    """Render unified non-CVE findings with asset, source, and evidence context."""
-    if not findings:
-        return ""
-
-    rows = []
-    for idx, finding in enumerate(sorted(findings, key=lambda f: severity_policy_rank(str(f.severity)), reverse=True)):
-        sev = str(finding.severity or "unknown").lower()
-        title = finding.title or finding.description or finding.finding_type.value
-        asset_label = finding.asset.name or finding.asset.identifier or "unknown asset"
-        location = (
-            f'<br><code style="color:#64748b;font-size:.72rem">{_esc(finding.asset.location)}</code>' if finding.asset.location else ""
-        )
-        description = finding.description or finding.remediation_guidance or ""
-        evidence_items = [f"{_esc(key)}={_esc(value)}" for key, value in finding.evidence.items() if value not in (None, "", [], {})][:4]
-        evidence = (
-            "<br>".join(f'<code style="color:#94a3b8;font-size:.72rem">{item}</code>' for item in evidence_items)
-            if evidence_items
-            else '<span style="color:#334155">&mdash;</span>'
-        )
-        remediation = finding.remediation_guidance or ""
-        description_html = f'<br><span style="color:#94a3b8">{_esc(description)}</span>' if description else ""
-        remediation_html = _esc(remediation) if remediation else '<span style="color:#334155">&mdash;</span>'
-        pg_cls = " pg-hidden" if idx >= _PAGE_SIZE else ""
-        rows.append(
-            f'<tr class="pg-row{pg_cls}" data-severity="{sev}" data-type="{_esc(finding.finding_type.value)}" '
-            f'data-source="{_esc(finding.source.value)}" data-asset-type="{_esc(finding.asset.asset_type)}">'
-            f"<td>{_sev_badge(sev)}</td>"
-            f'<td><code style="color:#c4b5fd;font-size:.75rem">{_esc(finding.finding_type.value)}</code></td>'
-            f'<td><strong style="color:#e2e8f0">{_esc(asset_label)}</strong>'
-            f'<br><span style="color:#64748b;font-size:.72rem">{_esc(finding.asset.asset_type)}</span>{location}</td>'
-            f'<td style="font-size:.82rem;color:#cbd5e1;max-width:260px"><strong>{_esc(title)}</strong>'
-            f"{description_html}</td>"
-            f'<td><code style="color:#94a3b8;font-size:.75rem">{_esc(finding.source.value)}</code></td>'
-            f'<td style="font-size:.75rem;color:#94a3b8;max-width:220px">{evidence}</td>'
-            f'<td style="font-size:.75rem;color:#4ade80;max-width:220px">{remediation_html}</td>'
-            "</tr>"
-        )
-
-    headers = ["Severity", "Type", "Asset", "Finding", "Source", "Evidence", "Remediation"]
-    header_html = "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-    severity_filters = "".join(
-        f'<label style="display:flex;align-items:center;gap:4px;font-size:.78rem;color:{color};cursor:pointer">'
-        f'<input type="checkbox" class="policy-sev-filter" value="{sev}" checked> {label}</label>'
-        for sev, label, color in (
-            ("critical", "Critical", "#fca5a5"),
-            ("high", "High", "#fb923c"),
-            ("medium", "Medium", "#fbbf24"),
-            ("low", "Low", "#94a3b8"),
-            ("unknown", "Unknown", "#94a3b8"),
-        )
-    )
-    finding_types = sorted({finding.finding_type.value for finding in findings})
-    type_options = "".join(f'<option value="{_esc(ftype)}">{_esc(ftype)}</option>' for ftype in finding_types)
-    asset_types = sorted({finding.asset.asset_type for finding in findings if finding.asset.asset_type})
-    asset_options = "".join(f'<option value="{_esc(asset_type)}">{_esc(asset_type)}</option>' for asset_type in asset_types)
-    filter_bar = (
-        '<div class="policy-filter-bar" style="display:flex;flex-wrap:wrap;gap:12px;align-items:center;'
-        'margin-bottom:14px;padding:12px 16px;background:#0f172a;border-radius:8px;border:1px solid #1e293b">'
-        '<span style="font-size:.72rem;color:#64748b;text-transform:uppercase;letter-spacing:.06em;font-weight:700">Filter:</span>'
-        f"{severity_filters}"
-        '<span style="width:1px;height:18px;background:#334155"></span>'
-        '<select id="policyTypeFilter" style="padding:6px 10px;background:#1e293b;border:1px solid #334155;border-radius:6px;'
-        'color:#e2e8f0;font-size:.78rem;outline:none"><option value="">All types</option>'
-        f"{type_options}</select>"
-        '<select id="policyAssetFilter" style="padding:6px 10px;background:#1e293b;border:1px solid #334155;border-radius:6px;'
-        'color:#e2e8f0;font-size:.78rem;outline:none"><option value="">All assets</option>'
-        f"{asset_options}</select>"
-        '<input type="text" id="policySearch" placeholder="Search policy findings&hellip;" '
-        'style="padding:6px 10px;background:#1e293b;border:1px solid #334155;border-radius:6px;'
-        'color:#e2e8f0;font-size:.78rem;width:220px;outline:none">'
-        '<span id="policyVisibleCount" style="margin-left:auto;font-size:.72rem;color:#64748b"></span>'
-        "</div>"
-    )
-    return (
-        f'<section id="policyfindings">'
-        f'<div class="sec-title">&#x1f6e1;&#xfe0f; Policy &amp; Security Findings'
-        f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">{len(findings)}</sup></div>'
-        f'<div class="panel">'
-        f'<div class="hint-box">'
-        f"Unified non-CVE findings from scanners, policy checks, runtime controls, and MCP intelligence. "
-        f"These are the same findings emitted in JSON and SARIF."
-        f"</div>"
-        f"{filter_bar}"
-        f'<div class="table-wrap"><table class="data-table sortable paginated" id="policyFindingsTable">'
-        f"<thead><tr>{header_html}</tr></thead>"
-        f"<tbody>{''.join(rows)}</tbody></table></div>"
-        f'{_pager_controls("policyFindingsTable", len(rows))}'
-        f"</div></section>"
-    )
-
-
-def _skill_audit_section(report: "AIBOMReport") -> str:
-    """Build the skill audit findings section if data is available."""
-    data = getattr(report, "skill_audit_data", None)
-    if not data:
-        return ""
-
-    findings = data.get("findings", [])
-    passed = data.get("passed", True)
-    pkgs_checked = data.get("packages_checked", 0)
-    servers_checked = data.get("servers_checked", 0)
-    creds_checked = data.get("credentials_checked", 0)
-    ai_summary = data.get("ai_skill_summary", "")
-    ai_risk = data.get("ai_overall_risk_level", "")
-
-    status_color = "#16a34a" if passed else "#dc2626"
-    status_text = "PASSED" if passed else "FAILED"
-
-    summary_html = ""
-    if ai_summary:
-        summary_html = (
-            f'<div class="hint-box" style="border-color:#818cf840;background:#1e1b4b40">'
-            f'<strong style="color:#c7d2fe">AI Analysis:</strong> {_esc(ai_summary)}'
-            f"</div>"
-        )
-
-    stats_html = (
-        f'<div style="display:flex;gap:24px;margin-bottom:16px;font-size:.82rem;color:#94a3b8">'
-        f'<span>Status: <strong style="color:{status_color}">{status_text}</strong></span>'
-        f"<span>Packages checked: <strong>{pkgs_checked}</strong></span>"
-        f"<span>Servers checked: <strong>{servers_checked}</strong></span>"
-        f"<span>Credentials checked: <strong>{creds_checked}</strong></span>"
-        + (
-            f'<span>AI risk level: <strong style="color:{_SEV_COLOR.get(ai_risk, "#64748b")}">{_esc(ai_risk).upper()}</strong></span>'
-            if ai_risk
-            else ""
-        )
-        + "</div>"
-    )
-
-    if not findings:
-        return (
-            f'<section id="skillaudit">'
-            f'<div class="sec-title">&#x1f6e1;&#xfe0f; Skill File Audit</div>'
-            f'<div class="panel">{stats_html}{summary_html}'
-            f'<div class="empty-state">&#x2705; No security findings in skill files.</div>'
-            f"</div></section>"
-        )
-
-    rows = []
-    for f in findings:
-        sev = f.get("severity", "low")
-        rows.append(
-            f"<tr>"
-            f"<td>{_sev_badge(sev)}</td>"
-            f'<td style="color:#e2e8f0;font-weight:600;font-size:.85rem">{_esc(f.get("title", ""))}</td>'
-            f'<td><code style="color:#94a3b8;font-size:.75rem">{_esc(f.get("category", ""))}</code></td>'
-            f'<td style="font-size:.78rem;color:#94a3b8;max-width:300px">{_esc(f.get("detail", ""))}</td>'
-            f'<td style="font-size:.75rem;color:#64748b">{_esc(f.get("source_file", ""))}</td>'
-            f'<td style="font-size:.75rem;color:#4ade80">{_esc(f.get("recommendation", ""))}</td>'
-            f"</tr>"
-        )
-
-    headers = ["Severity", "Finding", "Category", "Detail", "Source", "Recommendation"]
-    table_html = (
-        '<div class="table-wrap"><table class="data-table sortable">'
-        + "<thead><tr>"
-        + "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-        + "</tr></thead>"
-        + f"<tbody>{''.join(rows)}</tbody></table></div>"
-    )
-
-    return (
-        f'<section id="skillaudit">'
-        f'<div class="sec-title">&#x1f6e1;&#xfe0f; Skill File Audit'
-        f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">{len(findings)}</sup></div>'
-        f'<div class="panel">{stats_html}{summary_html}{table_html}</div>'
-        f"</section>"
-    )
-
-
-def _ai_inventory_section(report: "AIBOMReport") -> str:
-    """Build the AI component inventory section if data is available."""
-    data = getattr(report, "ai_inventory_data", None)
-    if not data:
-        return ""
-
-    components = data.get("components", [])
-    total = data.get("total_components", 0)
-    files_scanned = data.get("files_scanned", 0)
-    shadow_count = data.get("shadow_ai_count", 0)
-    deprecated_count = data.get("deprecated_models_count", 0)
-    unique_sdks = data.get("unique_sdks", [])
-    unique_models = data.get("unique_models", [])
-
-    stats_html = (
-        f'<div style="display:flex;gap:24px;margin-bottom:16px;font-size:.82rem;color:#94a3b8">'
-        f"<span>Files scanned: <strong>{files_scanned}</strong></span>"
-        f"<span>Components: <strong>{total}</strong></span>"
-        f"<span>SDKs: <strong>{len(unique_sdks)}</strong></span>"
-        f"<span>Models: <strong>{len(unique_models)}</strong></span>"
-        + (f'<span>Shadow AI: <strong style="color:#eab308">{shadow_count}</strong></span>' if shadow_count else "")
-        + (f'<span>Deprecated: <strong style="color:#f97316">{deprecated_count}</strong></span>' if deprecated_count else "")
-        + "</div>"
-    )
-
-    if not components:
-        return (
-            f'<section id="aiinventory">'
-            f'<div class="sec-title">&#x1f916; AI Component Inventory</div>'
-            f'<div class="panel">{stats_html}'
-            f'<div class="empty-state">&#x2705; No AI components detected.</div>'
-            f"</div></section>"
-        )
-
-    rows = []
-    for c in components:
-        sev = c.get("severity", "info")
-        comp_type = c.get("type", "").replace("_", " ")
-        # Redact API key values — never render credential fragments in HTML
-        raw_name = c.get("name", "")
-        name = "[REDACTED]" if c.get("type") == "api_key" else _esc(raw_name)
-        shadow_tag = ' <span style="color:#eab308;font-size:.7rem">(shadow)</span>' if c.get("is_shadow") else ""
-        replacement = c.get("deprecated_replacement", "")
-        repl_tag = f'<br><span style="color:#64748b;font-size:.7rem">&rarr; {_esc(replacement)}</span>' if replacement else ""
-        file_loc = f"{_esc(c.get('file', ''))}:{c.get('line', '')}"
-        rows.append(
-            f"<tr>"
-            f"<td>{_sev_badge(sev)}</td>"
-            f'<td><code style="color:#94a3b8;font-size:.75rem">{_esc(comp_type)}</code></td>'
-            f'<td style="color:#e2e8f0;font-weight:600;font-size:.85rem">{name}{shadow_tag}{repl_tag}</td>'
-            f'<td style="font-size:.75rem;color:#64748b">{file_loc}</td>'
-            f'<td style="font-size:.75rem;color:#67e8f9">{_esc(c.get("language", ""))}</td>'
-            f"</tr>"
-        )
-
-    headers = ["Severity", "Type", "Name", "File", "Language"]
-    table_html = (
-        '<div class="table-wrap"><table class="data-table sortable">'
-        + "<thead><tr>"
-        + "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-        + "</tr></thead>"
-        + f"<tbody>{''.join(rows)}</tbody></table></div>"
-    )
-
-    return (
-        f'<section id="aiinventory">'
-        f'<div class="sec-title">&#x1f916; AI Component Inventory'
-        f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">{total}</sup></div>'
-        f'<div class="panel">{stats_html}{table_html}</div>'
-        f"</section>"
-    )
-
-
-_CIS_CLOUD_LABELS = {
-    "aws": ("AWS", "cis_benchmark_data"),
-    "azure": ("Azure", "azure_cis_benchmark_data"),
-    "gcp": ("GCP", "gcp_cis_benchmark_data"),
-    "snowflake": ("Snowflake", "snowflake_cis_benchmark_data"),
-    "databricks": ("Databricks", "databricks_cis_benchmark_data"),
-}
-
-def _cis_evidence_html(check: dict) -> str:
-    """Affected-resource evidence cell: resource IDs when present, else text."""
-    resources = check.get("resource_ids") or []
-    if resources:
-        chips = "".join(
-            f'<span style="display:inline-block;padding:1px 7px;margin:1px 3px 1px 0;border-radius:4px;'
-            f"background:#0f172a;color:#cbd5e1;font-size:.66rem;font-family:monospace;"
-            f'word-break:break-all">{_esc(r)}</span>'
-            for r in resources[:6]
-        )
-        if len(resources) > 6:
-            chips += f'<span style="color:#475569;font-size:.66rem">+{len(resources) - 6} more</span>'
-        return chips
-    evidence = str(check.get("evidence") or "").strip()
-    if evidence:
-        return f'<span style="color:#94a3b8;font-size:.72rem">{_esc(evidence)}</span>'
-    return '<span style="color:#475569;font-size:.7rem">&mdash;</span>'
-
-
-def _cis_benchmark_section(report: "AIBOMReport") -> str:
-    """Build the CIS benchmark posture section (issue #665).
-
-    Renders one sub-panel per cloud with a CIS benchmark bundle. Each
-    failed check surfaces its structured remediation dict (``fix_cli``,
-    ``fix_console``, ``priority``, ``guardrails``, human-review flag).
-    Returns an empty string when no CIS data is present.
-    """
-
-    def _sort_key(c: dict) -> tuple[int, int, str]:
-        rem = c.get("remediation") or {}
-        priority = rem.get("priority", 3)
-        if not isinstance(priority, int):
-            priority = 3
-        return (severity_worst_first_rank(c.get("severity")), priority, str(c.get("check_id") or ""))
-
-    panels: list[str] = []
-    for idx, (cloud_key, (label, attr)) in enumerate(_CIS_CLOUD_LABELS.items()):
-        bundle = getattr(report, attr, None)
-        if not bundle:
-            continue
-        checks = bundle.get("checks") or []
-        if not checks:
-            continue
-        failed = [c for c in checks if c.get("status") == "fail"]
-        evaluated = [c for c in checks if c.get("status") in ("pass", "fail")]
-        passed_n = bundle.get("passed", sum(1 for c in checks if c.get("status") == "pass"))
-        pass_rate = bundle.get("pass_rate", 0.0)
-        band_color = "#16a34a" if pass_rate >= 90 else "#eab308" if pass_rate >= 70 else "#ef4444"
-
-        # Verdict + per-severity failed counts for the summary header.
-        sev_counts = {
-            s: sum(1 for c in failed if (c.get("severity") or "").lower() == s) for s in SEVERITY_THRESHOLD_LABELS
-        }
-        if not failed:
-            verdict_text, verdict_color = "PASS", "#16a34a"
-        else:
-            worst_check = min(failed, key=lambda c: severity_worst_first_rank(c.get("severity")))
-            worst_sev = (worst_check.get("severity") or "low").lower()
-            if worst_sev not in SEVERITY_THRESHOLD_LABELS:
-                worst_sev = "low"
-            worst_band = worst_sev.upper()
-            verdict_text = f"{worst_band} GAPS"
-            verdict_color = _SEV_COLOR.get(worst_sev, "#ef4444")
-
-        top = sorted(failed, key=_sort_key)[:3]
-        top_html = ""
-        if top:
-            chips = "".join(
-                f'<code style="background:#1e293b;color:#f1f5f9;padding:1px 6px;border-radius:4px;'
-                f'margin-right:5px;font-size:.7rem">{_esc(c.get("check_id", ""))}</code>'
-                for c in top
-            )
-            top_html = f'<div style="color:#64748b;font-size:.74rem;margin-top:6px">top risks: {chips}</div>'
-
-        counts_html = " ".join(
-            f'<span style="color:{_SEV_COLOR.get(s, "#6b7280")};font-size:.72rem;font-weight:700">{n} {s[0].upper()}</span>'
-            for s, n in sev_counts.items()
-            if n
-        )
-
-        header = (
-            f'<div style="display:flex;align-items:center;gap:18px;flex-wrap:wrap;margin-bottom:4px">'
-            f'<div style="font-weight:700;color:#f1f5f9;font-size:.95rem">{_esc(label)}</div>'
-            f'<span style="background:{verdict_color};color:#fff;padding:2px 9px;border-radius:4px;'
-            f'font-size:.7rem;font-weight:700;letter-spacing:.04em">{verdict_text}</span>'
-            f'<div style="color:{band_color};font-weight:700">{pass_rate:.0f}% pass</div>'
-            f'<div style="color:#64748b;font-size:.78rem">'
-            f"{passed_n}/{len(evaluated)} checks &middot; "
-            f'<strong style="color:#f97316">{len(failed)} failed</strong>'
-            f"</div>"
-            f'<div style="display:flex;gap:10px">{counts_html}</div>'
-            f"</div>"
-            f"{top_html}"
-        )
-
-        if not failed:
-            panels.append(
-                f'<div class="panel" style="margin-bottom:16px">{header}'
-                f'<div class="empty-state" style="margin-top:12px">&#x2705; No failed CIS checks.</div></div>'
-            )
-            continue
-
-        # Severity filter buttons (client-side, per panel).
-        panel_id = f"cis-{_esc(cloud_key)}"
-        filter_btns = (
-            f'<div class="cis-filter" data-target="{panel_id}" style="margin:10px 0 4px">'
-            + "".join(
-                f'<button type="button" class="cis-filter-btn" data-sev="{sev}" '
-                f'style="background:#1e293b;color:#94a3b8;border:1px solid #334155;border-radius:4px;'
-                f'padding:2px 9px;margin-right:6px;font-size:.7rem;cursor:pointer">{lbl}</button>'
-                for sev, lbl in (
-                    ("all", "All"),
-                    ("critical", "Critical"),
-                    ("high", "High"),
-                    ("medium", "Medium"),
-                    ("low", "Low"),
-                )
-                if sev == "all" or sev_counts.get(sev)
-            )
-            + "</div>"
-        )
-
-        rows = []
-        for check in sorted(failed, key=_sort_key):
-            sev = (check.get("severity") or "").lower()
-            rem = check.get("remediation") or {}
-            fix_cli = rem.get("fix_cli")
-            fix_console = rem.get("fix_console") or ""
-            effort = rem.get("effort") or "manual"
-            priority = rem.get("priority", 3)
-            guardrails = rem.get("guardrails") or []
-            guard_html = "".join(
-                f'<span style="display:inline-block;padding:1px 7px;margin:1px 3px 1px 0;border-radius:4px;background:#1e293b;color:#94a3b8;font-size:.65rem;font-family:monospace">{_esc(g)}</span>'
-                for g in guardrails[:5]
-            )
-            if len(guardrails) > 5:
-                guard_html += f'<span style="color:#475569;font-size:.65rem">+{len(guardrails) - 5}</span>'
-            review_badge = (
-                '<span style="color:#eab308;font-size:.7rem;margin-left:6px">&#8634; review</span>'
-                if rem.get("requires_human_review")
-                else ""
-            )
-
-            fix_cell = ""
-            if fix_cli:
-                fix_cell = f'<code style="color:#67e8f9;font-size:.72rem;white-space:pre-wrap;word-break:break-all">{_esc(fix_cli)}</code>'
-            elif fix_console:
-                fix_cell = f'<span style="color:#94a3b8;font-size:.72rem">&rarr; {_esc(fix_console)}</span>'
-            elif check.get("recommendation"):
-                fix_cell = f'<span style="color:#94a3b8;font-size:.72rem">{_esc(check.get("recommendation"))}</span>'
-
-            docs_link = ""
-            docs = rem.get("docs") or ""
-            if docs:
-                docs_link = f' &middot; <a href="{_esc(docs)}" target="_blank" rel="noopener" style="font-size:.7rem">docs</a>'
-
-            rows.append(
-                f'<tr class="cis-row" data-sev="{_esc(sev)}">'
-                f"<td>{_sev_badge(sev)}</td>"
-                f'<td style="font-family:monospace;color:#f1f5f9;font-weight:600">{_esc(check.get("check_id", ""))}</td>'
-                f'<td style="color:#e2e8f0;font-size:.82rem">{_esc(check.get("title", ""))}{review_badge}</td>'
-                f'<td style="color:#94a3b8;font-size:.75rem">P{priority} &middot; {_esc(effort)}</td>'
-                f"<td>{_cis_evidence_html(check)}</td>"
-                f"<td>{guard_html}</td>"
-                f"<td>{fix_cell}{docs_link}</td>"
-                "</tr>"
-            )
-
-        table_html = (
-            f'<div class="table-wrap"><table class="data-table sortable" id="{panel_id}">'
-            + "<thead><tr>"
-            + "".join(
-                f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>'
-                for i, h in enumerate(["Severity", "Check", "Title", "Priority", "Evidence", "Guardrails", "Remediation"])
-            )
-            + "</tr></thead>"
-            + f"<tbody>{''.join(rows)}</tbody></table></div>"
-        )
-
-        panels.append(f'<div class="panel" style="margin-bottom:16px">{header}{filter_btns}{table_html}</div>')
-
-    if not panels:
-        return ""
-
-    filter_script = (
-        "<script>document.querySelectorAll('.cis-filter').forEach(function(bar){"
-        "bar.querySelectorAll('.cis-filter-btn').forEach(function(btn){"
-        "btn.addEventListener('click',function(){"
-        "var tbl=document.getElementById(bar.getAttribute('data-target'));if(!tbl)return;"
-        "var sev=btn.getAttribute('data-sev');"
-        "tbl.querySelectorAll('tbody tr.cis-row').forEach(function(r){"
-        "r.style.display=(sev==='all'||r.getAttribute('data-sev')===sev)?'':'none';});"
-        "bar.querySelectorAll('.cis-filter-btn').forEach(function(b){b.style.color='#94a3b8';});"
-        "btn.style.color='#f1f5f9';});});});</script>"
-    )
-
-    return (
-        '<section id="cisbenchmarks"><div class="sec-title">&#x1f6e1;&#xfe0f; CIS Benchmark Posture</div>'
-        + "".join(panels)
-        + filter_script
-        + "</section>"
-    )
-
-
-def _trust_assessment_section(report: "AIBOMReport") -> str:
-    """Build the trust assessment section if data is available."""
-    data = getattr(report, "trust_assessment_data", None)
-    if not data:
-        return ""
-
-    skill_name = _esc(data.get("skill_name", ""))
-    source_file = _esc(data.get("source_file", ""))
-    verdict = data.get("verdict", "benign").upper()
-    content_verdict = data.get("content_verdict", data.get("verdict", "benign")).upper()
-    provenance_verdict = data.get("provenance_verdict", "unverified").upper()
-    recommendation = data.get("overall_recommendation") or data.get("review_verdict", "review")
-    confidence = data.get("confidence", "low")
-    categories = data.get("categories", [])
-    recommendations = data.get("recommendations", [])
-
-    verdict_colors = {"BENIGN": "#16a34a", "SUSPICIOUS": "#d97706", "MALICIOUS": "#dc2626"}
-    verdict_color = verdict_colors.get(verdict, "#6b7280")
-
-    level_icons = {"pass": "&#x2713;", "info": "&#x2139;", "warn": "&#x26a0;", "fail": "&#x2717;"}
-    level_colors = {"pass": "#16a34a", "info": "#60a5fa", "warn": "#d97706", "fail": "#dc2626"}
-
-    cat_rows = []
-    for cat in categories:
-        level = cat.get("level", "pass")
-        icon = level_icons.get(level, "?")
-        color = level_colors.get(level, "#6b7280")
-        cat_rows.append(
-            f"<tr>"
-            f'<td style="text-align:center;color:{color};font-size:1.1rem">{icon}</td>'
-            f'<td style="color:#e2e8f0;font-weight:600;font-size:.85rem">{_esc(cat.get("name", ""))}</td>'
-            f"<td>{_sev_badge(level)}</td>"
-            f'<td style="font-size:.78rem;color:#94a3b8">{_esc(cat.get("summary", ""))}</td>'
-            f"</tr>"
-        )
-
-    verdict_badge = (
-        f'<span style="background:{verdict_color};color:#fff;padding:4px 14px;border-radius:6px;'
-        f'font-size:.82rem;font-weight:700;letter-spacing:.04em">{verdict}</span>'
-        f'<span style="color:#64748b;font-size:.78rem;margin-left:8px">({confidence} confidence)</span>'
-    )
-    axis_html = (
-        '<div style="display:flex;gap:8px;flex-wrap:wrap;margin-bottom:16px">'
-        f'<span style="background:#0f766e;color:#fff;padding:4px 10px;border-radius:6px;font-size:.76rem">'
-        f"content: {_esc(content_verdict)}</span>"
-        f'<span style="background:#4338ca;color:#fff;padding:4px 10px;border-radius:6px;font-size:.76rem">'
-        f"provenance: {_esc(provenance_verdict)}</span>"
-        f'<span style="background:#475569;color:#fff;padding:4px 10px;border-radius:6px;font-size:.76rem">'
-        f"recommendation: {_esc(str(recommendation).upper())}</span>"
-        "</div>"
-    )
-
-    title_suffix = f" &mdash; {skill_name}" if skill_name else ""
-    source_note = (
-        f'<div style="font-size:.72rem;color:#475569;margin-bottom:12px">Source: <code>{source_file}</code></div>' if source_file else ""
-    )
-
-    rec_html = ""
-    if recommendations:
-        rec_items = "".join(f'<li style="color:#4ade80;font-size:.78rem;margin-bottom:4px">{_esc(r)}</li>' for r in recommendations)
-        rec_html = f'<ul style="list-style:none;padding:0;margin-top:16px">{rec_items}</ul>'
-
-    headers = ["", "Category", "Level", "Summary"]
-    table_html = (
-        '<div class="table-wrap"><table class="data-table">'
-        + "<thead><tr>"
-        + "".join(f"<th>{h}</th>" for h in headers)
-        + "</tr></thead>"
-        + f"<tbody>{''.join(cat_rows)}</tbody></table></div>"
-    )
-
-    return (
-        f'<section id="trust">'
-        f'<div class="sec-title">&#x1f50d; Trust Assessment{title_suffix}</div>'
-        f'<div class="panel">'
-        f"{source_note}"
-        f'<div style="margin-bottom:16px">{verdict_badge}</div>'
-        f"{axis_html}"
-        f"{table_html}"
-        f"{rec_html}"
-        f"</div></section>"
-    )
-
-
-def _enforcement_section(report: "AIBOMReport") -> str:
-    """Build the enforcement findings section if data is available."""
-    data = getattr(report, "enforcement_data", None)
-    if not data:
-        return ""
-
-    findings = data.get("findings", [])
-    passed = data.get("passed", True)
-    servers_checked = data.get("servers_checked", 0)
-    tools_checked = data.get("tools_checked", 0)
-    critical_count = data.get("critical_count", 0)
-    high_count = data.get("high_count", 0)
-
-    status_color = "#16a34a" if passed else "#dc2626"
-    status_text = "PASSED" if passed else "FAILED"
-
-    stats_html = (
-        f'<div style="display:flex;gap:24px;margin-bottom:16px;font-size:.82rem;color:#94a3b8">'
-        f'<span>Status: <strong style="color:{status_color}">{status_text}</strong></span>'
-        f"<span>Servers checked: <strong>{servers_checked}</strong></span>"
-        f"<span>Tools checked: <strong>{tools_checked}</strong></span>"
-        f'<span>Critical: <strong style="color:#dc2626">{critical_count}</strong></span>'
-        f'<span>High: <strong style="color:#ea580c">{high_count}</strong></span>'
-        f"</div>"
-    )
-
-    if not findings:
-        return (
-            f'<section id="enforcement">'
-            f'<div class="sec-title">&#x1f6a8; Enforcement</div>'
-            f'<div class="panel">{stats_html}'
-            f'<div class="empty-state">&#x2705; No enforcement findings — all checks passed.</div>'
-            f"</div></section>"
-        )
-
-    rows = []
-    for f in findings:
-        sev = f.get("severity", "low")
-        rows.append(
-            f"<tr>"
-            f"<td>{_sev_badge(sev)}</td>"
-            f'<td><code style="color:#94a3b8;font-size:.75rem">{_esc(f.get("category", ""))}</code></td>'
-            f'<td style="color:#e2e8f0;font-weight:600;font-size:.85rem">{_esc(f.get("server_name", ""))}</td>'
-            f'<td style="font-size:.78rem;color:#94a3b8">{_esc(f.get("tool_name", "") or "—")}</td>'
-            f'<td style="font-size:.78rem;color:#94a3b8;max-width:350px">{_esc(f.get("reason", ""))}</td>'
-            f'<td style="font-size:.75rem;color:#4ade80">{_esc(f.get("recommendation", ""))}</td>'
-            f"</tr>"
-        )
-
-    headers = ["Severity", "Category", "Server", "Tool", "Reason", "Recommendation"]
-    table_html = (
-        '<div class="table-wrap"><table class="data-table sortable">'
-        + "<thead><tr>"
-        + "".join(f'<th data-col="{i}">{h} <span class="sort-arrow"></span></th>' for i, h in enumerate(headers))
-        + "</tr></thead>"
-        + f"<tbody>{''.join(rows)}</tbody></table></div>"
-    )
-
-    return (
-        f'<section id="enforcement">'
-        f'<div class="sec-title">&#x1f6a8; Enforcement'
-        f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">{len(findings)}</sup></div>'
-        f'<div class="panel">{stats_html}{table_html}</div>'
-        f"</section>"
-    )
-
-
-def _attack_flow_section(findings: list["Finding"]) -> str:
-    """Build the CVE attack flow graph section (only when vulns exist)."""
-    if not findings:
-        return ""
-
-    total_creds = len({cred for finding in findings for cred in finding.exposed_credentials})
-    total_tools = len({tool for finding in findings for tool in finding.exposed_tools})
-    total_agents = len({agent for finding in findings for agent in finding.affected_agents})
-
-    return (
-        '<section id="attackflow">'
-        '<div class="sec-title">&#x1f525; CVE Attack Flow'
-        '<span style="font-size:.68rem;font-weight:400;opacity:.5;margin-left:8px">'
-        f"{len(findings)} CVEs &#x2192; {total_agents} agents &#x2192; "
-        f"{total_creds} credentials &#x2192; {total_tools} tools at risk"
-        "</span></div>"
-        '<div class="graph-container">'
-        '<div id="cyAttack" class="cy-graph"></div>'
-        '<div class="graph-controls" style="top:12px;right:12px">'
-        '<button class="graph-btn" id="afZoomIn" title="Zoom in">+</button>'
-        '<button class="graph-btn" id="afZoomOut" title="Zoom out">&minus;</button>'
-        '<button class="graph-btn" id="afFitBtn" title="Fit to view">&#x2922;</button>'
-        "</div>"
-        "</div>"
-        '<div class="legend">'
-        '<span><i class="diamond" style="background:#f87171"></i>CVE</span>'
-        '<span><i style="background:#dc2626"></i>Vulnerable Package</span>'
-        '<span><i style="background:#475569"></i>MCP Server</span>'
-        '<span><i style="background:#fbbf24"></i>Credential</span>'
-        '<span><i style="background:#818cf8"></i>Tool</span>'
-        '<span><i style="background:#3b82f6"></i>Agent</span>'
-        "</div>"
-        "</section>"
-    )
-
-
-def _inventory_cards(report: "AIBOMReport") -> str:
-    cards = []
-    for agent in report.agents:
-        total_vulns = agent.total_vulnerabilities
-        total_creds = sum(len(s.credential_names) for s in agent.mcp_servers)
-        agent_badges = []
-        if total_vulns:
-            agent_badges.append(f'<span class="badge-vuln">{total_vulns} vuln{"s" if total_vulns != 1 else ""}</span>')
-        if total_creds:
-            agent_badges.append(f'<span class="badge-cred">{total_creds} credential{"s" if total_creds != 1 else ""}</span>')
-        badges_html = " ".join(agent_badges)
-
-        servers_html = []
-        for srv in agent.mcp_servers:
-            vuln_pkgs = [p for p in srv.packages if p.has_vulnerabilities]
-            accent = "#ef4444" if vuln_pkgs else ("#f59e0b" if srv.has_credentials else "#334155")
-            srv_badges = []
-            if vuln_pkgs:
-                srv_badges.append('<span class="badge-vuln">VULN</span>')
-            if srv.has_credentials:
-                srv_badges.append('<span class="badge-cred">CREDS</span>')
-
-            cmd = ""
-            if srv.command:
-                cmd = _esc(sanitize_launch_command(srv.command, srv.args, max_args=3))
-                if srv.args and len(srv.args) > 3:
-                    cmd += f' <span style="color:#334155">&hellip;+{len(srv.args) - 3} args</span>'
-
-            # Credentials section
-            creds_html = ""
-            if srv.credential_names:
-                creds_html = (
-                    '<div style="margin-top:8px">'
-                    + "".join(
-                        f'<div style="font-size:.74rem;color:#fbbf24;padding:2px 0">&#x1f511; <code>{_esc(c)}</code></div>'
-                        for c in srv.credential_names
-                    )
-                    + "</div>"
-                )
-
-            # Packages — preview first N, collapse rest
-            pkgs = srv.packages
-            pkg_count = len(pkgs)
-            preview = pkgs[:_PKG_PREVIEW]
-            rest = pkgs[_PKG_PREVIEW:]
-
-            def pkg_row(p: object) -> str:
-                from agent_bom.models import Package
-
-                if not isinstance(p, Package):
-                    return ""
-                color = "#f87171" if p.has_vulnerabilities else "#38bdf8"
-                vuln_mark = " &#x26a0;" if p.has_vulnerabilities else ""
-                return (
-                    f'<div class="pkg-row">'
-                    f'<span><code style="color:{color};font-size:.72rem">{_esc(p.ecosystem)}</code>'
-                    f' <span class="pkg-name">{_esc(p.name)}</span></span>'
-                    f'<span class="pkg-ver">{_esc(p.version)}{vuln_mark}</span>'
-                    f"</div>"
-                )
-
-            pkg_html = ""
-            if pkgs:
-                preview_rows = "".join(pkg_row(p) for p in preview)
-                if rest:
-                    uid = f"pkgs_{id(srv)}"
-                    rest_rows = "".join(pkg_row(p) for p in rest)
-                    pkg_html = (
-                        f'<div style="margin-top:8px">'
-                        f"{preview_rows}"
-                        f'<div id="{uid}" style="display:none">{rest_rows}</div>'
-                        f'<button class="toggle-btn" onclick="togglePkgs(\'{uid}\',this)">'
-                        f"Show {len(rest)} more packages &#x25bc;</button>"
-                        f"</div>"
-                    )
-                else:
-                    pkg_html = f'<div style="margin-top:8px">{preview_rows}</div>'
-
-            srv_badges_html = " ".join(srv_badges)
-            srv_header = (
-                f'<div style="display:flex;justify-content:space-between;align-items:center;margin-bottom:6px">'
-                f'<div style="font-weight:600;color:#e2e8f0;font-size:.9rem">&#x2699;&#xfe0f; {_esc(srv.name)} {srv_badges_html}</div>'
-                f'<div style="font-size:.72rem;color:#475569">{pkg_count} pkg{"s" if pkg_count != 1 else ""}</div>'
-                f"</div>"
-            )
-            cmd_html = (
-                f'<div style="font-size:.72rem;color:#475569;font-family:monospace;margin-bottom:6px;word-break:break-all">{cmd}</div>'
-                if cmd
-                else ""
-            )
-
-            servers_html.append(
-                f'<div class="server-card" style="border-left-color:{accent}">{srv_header}{cmd_html}{creds_html}{pkg_html}</div>'
-            )
-
-        servers_content = (
-            "".join(servers_html) if servers_html else ('<p style="color:#334155;font-size:.85rem">No MCP servers configured.</p>')
-        )
-
-        cards.append(
-            f'<details class="agent-card" open>'
-            f'<summary class="agent-summary">'
-            f"<span>&#x1f916; {_esc(agent.name)}</span>"
-            f'<span style="display:flex;align-items:center;gap:8px">'
-            f"{badges_html}"
-            f'<span style="font-size:.72rem;color:#475569">'
-            f"{len(agent.mcp_servers)} server(s) &middot; {agent.total_packages} pkg(s)"
-            f"</span>"
-            f"</span>"
-            f"</summary>"
-            f'<div class="agent-detail">'
-            f'<div style="font-size:.72rem;color:#475569;margin-bottom:12px">'
-            f"{_esc(agent.agent_type.value)} &middot; {_esc(sanitize_path_label(agent.config_path) if agent.config_path else '')}"
-            f"</div>"
-            f"{servers_content}"
-            f"</div>"
-            f"</details>"
-        )
-    return "".join(cards)
-
-
-# ─── Compliance section ──────────────────────────────────────────────────────
-
-_STATUS_BADGE = {
-    "pass": '<span style="background:#16a34a;color:#fff;padding:2px 8px;border-radius:4px;font-size:.75rem">PASS</span>',
-    "warning": '<span style="background:#d97706;color:#fff;padding:2px 8px;border-radius:4px;font-size:.75rem">WARN</span>',
-    "fail": '<span style="background:#dc2626;color:#fff;padding:2px 8px;border-radius:4px;font-size:.75rem">FAIL</span>',
-}
-
-
-def _compliance_section(findings: list["Finding"]) -> str:
-    """Build OWASP/ATLAS/ATT&CK/NIST compliance tables from unified CVE findings."""
-    try:
-        from agent_bom.atlas import ATLAS_TECHNIQUES
-        from agent_bom.eu_ai_act import EU_AI_ACT
-        from agent_bom.mitre_attack import ATTACK_TECHNIQUES
-        from agent_bom.nist_ai_rmf import NIST_AI_RMF
-        from agent_bom.owasp import OWASP_LLM_TOP10
-        from agent_bom.owasp_agentic import OWASP_AGENTIC_TOP10
-    except ImportError:
-        return ""
-
-    br_dicts = [compliance_row_dict(finding) for finding in findings]
-
-    def _build_rows(catalog: dict[str, str], tag_field: str) -> tuple[str, int, int, int]:
-        rows = []
-        pass_count = fail_count = warn_count = 0
-        for code, name in sorted(catalog.items()):
-            findings = sum(1 for b in br_dicts if code in b.get(tag_field, []))
-            sev_counts = {"critical": 0, "high": 0, "medium": 0, "low": 0}
-            for b in br_dicts:
-                if code in b.get(tag_field, []):
-                    s = b.get("severity", "").lower()
-                    if s in sev_counts:
-                        sev_counts[s] += 1
-            if findings == 0:
-                status = "pass"
-                pass_count += 1
-            elif sev_counts["critical"] > 0 or sev_counts["high"] > 0:
-                status = "fail"
-                fail_count += 1
-            else:
-                status = "warning"
-                warn_count += 1
-            badge = _STATUS_BADGE[status]
-            rows.append(
-                f"<tr><td><code>{_esc(code)}</code></td><td>{_esc(name)}</td>"
-                f"<td style='text-align:center'>{findings}</td><td style='text-align:center'>{badge}</td></tr>"
-            )
-        return "\n".join(rows), pass_count, fail_count, warn_count
-
-    owasp_rows, op, of, ow = _build_rows(OWASP_LLM_TOP10, "owasp_tags")
-    atlas_rows, ap, af, aw = _build_rows(ATLAS_TECHNIQUES, "atlas_tags")
-    attack_rows, atp, atf, atw = _build_rows(dict(ATTACK_TECHNIQUES), "attack_tags")
-    nist_rows, np_, nf, nw = _build_rows(NIST_AI_RMF, "nist_ai_rmf_tags")
-    agentic_rows, oap, oaf, oaw = _build_rows(OWASP_AGENTIC_TOP10, "owasp_agentic_tags")
-    eu_rows, ep, ef_, ew = _build_rows(EU_AI_ACT, "eu_ai_act_tags")
-
-    total = (op + of + ow) + (ap + af + aw) + (atp + atf + atw) + (np_ + nf + nw) + (oap + oaf + oaw) + (ep + ef_ + ew)
-    total_pass = op + ap + atp + np_ + oap + ep
-    score = round((total_pass / total) * 100, 1) if total > 0 else 0.0
-    has_fail = (of + af + atf + nf + oaf + ef_) > 0
-    has_warn = (ow + aw + atw + nw + oaw + ew) > 0
-    overall = "fail" if has_fail else ("warning" if has_warn else "pass")
-    overall_badge = _STATUS_BADGE[overall]
-
-    def _framework_table(title: str, rows: str, p: int, f: int, w: int) -> str:
-        sub_total = p + f + w
-        return (
-            f'<details style="margin-bottom:12px" {"open" if f > 0 else ""}>'
-            f'<summary style="cursor:pointer;font-weight:600;padding:6px 0">'
-            f'{title} <span style="font-size:.8rem;color:#94a3b8">({p}/{sub_total} pass)</span></summary>'
-            f'<table class="vtable" style="margin-top:8px"><thead><tr>'
-            f"<th>Code</th><th>Control</th><th>Findings</th><th>Status</th></tr></thead>"
-            f"<tbody>{rows}</tbody></table></details>"
-        )
-
-    return (
-        f'<section id="compliance">'
-        f'<div class="sec-title">&#x1f6e1;&#xfe0f; Compliance Posture'
-        f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">'
-        f"Score: {score}% {overall_badge}</sup></div>"
-        f'<div class="panel">'
-        f"{_framework_table('OWASP LLM Top 10', owasp_rows, op, of, ow)}"
-        f"{_framework_table('OWASP Agentic Top 10', agentic_rows, oap, oaf, oaw)}"
-        f"{_framework_table('MITRE ATT&CK Enterprise', attack_rows, atp, atf, atw)}"
-        f"{_framework_table('MITRE ATLAS (AI/ML)', atlas_rows, ap, af, aw)}"
-        f"{_framework_table('NIST AI RMF', nist_rows, np_, nf, nw)}"
-        f"{_framework_table('EU AI Act', eu_rows, ep, ef_, ew)}"
-        f"</div></section>"
-    )
-
-
-# ─── Main assembler ───────────────────────────────────────────────────────────
-
-
-def _offline_assets_notice() -> str:
-    return """
-  <div class="offline-assets-banner" style="margin-bottom:16px;padding:12px 14px;border:1px solid #334155;border-radius:10px;background:#111827;color:#cbd5e1">
-    <strong style="color:#f8fafc">Offline HTML mode</strong>
-    <span style="color:#94a3b8"> — external JavaScript assets were omitted. Static tables, findings, remediation, compliance, inventory, and evidence sections remain available.</span>
-  </div>
-"""
-
-
-def _offline_assets_script() -> str:
-    return """<script>
-(function() {
-  function replace(id, title) {
-    var el = document.getElementById(id);
-    if (!el) return;
-    var box = document.createElement('div');
-    box.style.cssText = 'min-height:180px;display:flex;align-items:center;justify-content:center;text-align:center;padding:24px;border:1px dashed #334155;border-radius:10px;color:#94a3b8;background:#0f172a';
-    box.innerHTML = '<div><strong style="color:#cbd5e1">' + title + '</strong><br>Interactive rendering is disabled in offline HTML mode.</div>';
-    el.parentNode.replaceChild(box, el);
-  }
-  replace('sevChart', 'Severity chart');
-  replace('blastChart', 'Blast-radius chart');
-  replace('cy', 'Supply-chain graph');
-  replace('attackCy', 'Attack-flow graph');
-  document.querySelectorAll('.graph-filter-bar,.graph-controls').forEach(function(el) {
-    el.style.display = 'none';
-  });
-  document.querySelectorAll('details').forEach(function(el) {
-    el.open = el.open || false;
-  });
-})();
-</script>"""
-
-
-def _apply_offline_assets_mode(html: str) -> str:
-    for tag in _EXTERNAL_SCRIPT_TAGS:
-        html = html.replace(tag, "")
-    marker = '<div class="container">\n'
-    html = html.replace(marker, marker + _offline_assets_notice(), 1)
-    script_start = html.find("\n<script>\n(function() {")
-    script_end = html.rfind("</script>\n\n</body>")
-    if script_start != -1 and script_end != -1:
-        html = html[:script_start] + "\n" + _offline_assets_script() + html[script_end + len("</script>") :]
-    return html
-
-
-def to_html(
-    report: "AIBOMReport",
-    blast_radii: list["BlastRadius"] | None = None,
-    *,
-    offline_assets: bool = False,
-) -> str:
-    blast_radii = blast_radii or []
-    findings = cve_findings(report, blast_radii)
-    policy_findings = _non_cve_findings(report)
-    generated = report.generated_at.strftime("%Y-%m-%d %H:%M:%S UTC")
-    elements_json = _cytoscape_elements(report, blast_radii)
-    attack_flow_json = _attack_flow_elements(report, blast_radii)
-    chart_data_json = _chart_data(findings)
-    crit = sum(1 for finding in findings if severity_value(finding) == "critical")
-    policy_crit = sum(1 for finding in policy_findings if str(finding.severity).lower() == "critical")
-    policy_high = sum(1 for finding in policy_findings if str(finding.severity).lower() == "high")
-    total_vulns = len(findings)
-
-    if crit or policy_crit:
-        status_color, status_label = "#dc2626", "CRITICAL FINDINGS"
-    elif total_vulns or policy_high or policy_findings:
-        status_color, status_label = "#d97706", "SECURITY FINDINGS"
-    else:
-        status_color, status_label = "#16a34a", "CLEAN"
-
-    # Sections
-    vuln_sections = ""
-    if findings:
-        vuln_sections = (
-            f'<section id="vulns">'
-            f'<div class="sec-title">&#x26a0;&#xfe0f; Vulnerabilities'
-            f'<sup style="font-size:.7rem;color:#475569;margin-left:6px">{len(findings)}</sup>'
-            f"</div>"
-            f'<div class="panel">{_vuln_table(report, blast_radii)}</div>'
-            f"</section>"
-            f'<section id="exposure-paths">'
-            f'<div class="sec-title">&#x1f9ed; Exposure Paths'
-            f'<sup style="font-size:.65rem;color:#475569;margin-left:6px;font-weight:400">'
-            f"ranked investigation briefs from blast-radius evidence"
-            f"</sup></div>"
-            f'<div class="panel exposure-paths">{_exposure_path_section(report, blast_radii)}</div>'
-            f"</section>"
-            f'<section id="blast">'
-            f'<div class="sec-title">&#x1f4a5; Blast Radius'
-            f'<sup style="font-size:.65rem;color:#475569;margin-left:6px;font-weight:400">'
-            f"risk = CVSS + agents + creds + tools + KEV/EPSS boosts (max 10)"
-            f"</sup></div>"
-            f'<div class="panel">{_blast_table(report, blast_radii)}</div>'
-            f"</section>"
-            f'<section id="remediation">'
-            f'<div class="sec-title">&#x1f527; Remediation Plan</div>'
-            f'<div class="panel">{_remediation_list(findings)}</div>'
-            f"</section>"
-        )
-    policy_section = _policy_findings_section(policy_findings)
-
-    # Compliance section
-    compliance_html = _compliance_section(findings)
-
-    # AI inventory section
-    ai_inv_section = _ai_inventory_section(report)
-
-    # Skill audit section
-    skill_section = _skill_audit_section(report)
-
-    # Trust assessment section
-    trust_section = _trust_assessment_section(report)
-
-    # Enforcement section
-    enforce_section = _enforcement_section(report)
-
-    # CIS benchmark posture (issue #665 — structured remediation)
-    cis_bench_section = _cis_benchmark_section(report)
-
-    # Determine node counts for graph subtitle
-    vuln_node_count = len({topology_package_key(finding) for finding in findings})
-    graph_note = (
-        f"agents + servers + {vuln_node_count} vulnerable pkg(s) only — {report.total_packages - vuln_node_count} clean packages hidden"
-        if report.total_packages > vuln_node_count
-        else "agents + servers + packages"
-    )
-
-    html = f"""<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width,initial-scale=1">
-  <title>agent-bom AI-BOM &mdash; {_esc(generated)}</title>
-  <script src="https://cdn.jsdelivr.net/npm/chart.js@4.4.2/dist/chart.umd.min.js"></script>
-  <script src="https://unpkg.com/cytoscape@3.30.2/dist/cytoscape.min.js"></script>
-  <script src="https://unpkg.com/dagre@0.8.5/dist/dagre.min.js"></script>
-  <script src="https://unpkg.com/cytoscape-dagre@2.5.0/cytoscape-dagre.js"></script>
-  <script src="https://unpkg.com/cytoscape-popper@2.0.0/cytoscape-popper.js"></script>
-  <style>
-    *,*::before,*::after{{box-sizing:border-box;margin:0;padding:0}}
-    body{{font-family:-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,sans-serif;background:#0b1120;color:#cbd5e1;line-height:1.6;font-size:14px}}
-    a{{color:#60a5fa;text-decoration:none}}
-    a:hover{{text-decoration:underline}}
-    code{{font-family:"SF Mono","Cascadia Code",Consolas,monospace;font-size:.9em}}
-
-    /* SIDEBAR NAV */
-    .sidebar{{position:fixed;top:0;left:0;bottom:0;width:220px;background:#0a0f1a;border-right:1px solid #1e293b;z-index:100;display:flex;flex-direction:column;overflow-y:auto;transition:width .2s}}
-    .sidebar-brand{{display:flex;align-items:center;gap:10px;padding:18px 18px 14px;border-bottom:1px solid #1e293b18}}
-    .sidebar-brand .brand-icon{{width:32px;height:32px;border-radius:8px;background:rgba(16,185,129,.1);border:1px solid rgba(16,185,129,.2);display:flex;align-items:center;justify-content:center;font-size:1rem}}
-    .sidebar-brand .brand-text{{font-weight:700;font-size:.88rem;color:#f1f5f9;letter-spacing:-.01em}}
-    .sidebar-brand .brand-sub{{font-size:.6rem;color:#475569;font-family:monospace}}
-    .sidebar-status{{padding:8px 18px 12px;display:flex;align-items:center;gap:8px}}
-    .status-badge{{padding:3px 10px;border-radius:5px;font-size:.62rem;font-weight:700;letter-spacing:.05em;background:{status_color}12;color:{status_color};border:1px solid {status_color}25;white-space:nowrap}}
-    .scan-time{{color:#475569;font-size:.62rem;white-space:nowrap}}
-    .sidebar-group{{padding:4px 0}}
-    .sidebar-group-label{{font-size:.58rem;font-weight:700;letter-spacing:.1em;text-transform:uppercase;color:#3f4f65;padding:10px 18px 4px;user-select:none}}
-    .sidebar-link{{display:flex;align-items:center;gap:8px;padding:7px 18px;font-size:.78rem;color:#64748b;text-decoration:none;border-left:2px solid transparent;transition:all .12s;cursor:pointer}}
-    .sidebar-link:hover{{background:#111827;color:#e2e8f0;text-decoration:none;border-left-color:#334155}}
-    .sidebar-link.active{{background:rgba(16,185,129,.08);color:#10b981;border-left-color:#10b981;font-weight:600}}
-    .sidebar-link .link-icon{{font-size:.85rem;width:20px;text-align:center;flex-shrink:0}}
-    .sidebar-link .link-badge{{margin-left:auto;font-size:.6rem;font-weight:700;padding:1px 6px;border-radius:10px;font-family:monospace}}
-    .sidebar-spacer{{flex:1}}
-    .sidebar-footer{{padding:14px 18px;border-top:1px solid #1e293b18;font-size:.65rem;color:#334155}}
-    .sidebar-footer a{{color:#475569}}
-    .print-btn{{background:transparent;border:1px solid #1e293b;color:#475569;font-size:.68rem;padding:5px 12px;border-radius:6px;cursor:pointer;transition:all .15s;width:100%;margin-top:8px;text-align:center}}
-    .print-btn:hover{{background:#111827;color:#94a3b8}}
-
-    /* MOBILE SIDEBAR TOGGLE */
-    .sidebar-toggle{{display:none;position:fixed;top:12px;left:12px;z-index:200;width:36px;height:36px;border-radius:8px;background:rgba(10,15,26,.95);border:1px solid #1e293b;color:#94a3b8;font-size:18px;cursor:pointer;backdrop-filter:blur(8px)}}
-
-    /* LAYOUT */
-    .container{{max-width:1400px;margin:0 auto;padding:28px 32px 80px;margin-left:220px}}
-    section{{margin-bottom:44px;scroll-margin-top:20px}}
-    .sec-title{{font-size:.82rem;font-weight:700;letter-spacing:.08em;text-transform:uppercase;color:#64748b;margin-bottom:18px;padding-bottom:10px;border-bottom:1px solid #1e293b}}
-    .panel{{background:#1e293b;border-radius:12px;padding:24px;border:1px solid #ffffff08}}
-
-    /* STAT CARDS */
-    .stat-grid{{display:grid;grid-template-columns:repeat(auto-fill,minmax(180px,1fr));gap:14px}}
-    .stat-card{{background:linear-gradient(135deg,#1e293b 0%,#0f172a 100%);border-radius:10px;padding:20px 22px;border-left:4px solid #334155;border:1px solid #ffffff06;transition:transform .15s,box-shadow .15s}}
-    .stat-card:hover{{transform:translateY(-2px);box-shadow:0 8px 24px rgba(0,0,0,.3)}}
-    .stat-icon{{font-size:1.4rem;margin-bottom:6px}}
-    .stat-value{{font-size:2.2rem;font-weight:800;line-height:1;margin-bottom:6px}}
-    .stat-label{{font-size:.7rem;color:#64748b;text-transform:uppercase;letter-spacing:.06em}}
-
-    /* CHARTS ROW */
-    .charts-row{{display:grid;grid-template-columns:320px 1fr;gap:16px}}
-    .chart-panel{{background:linear-gradient(135deg,#1e293b 0%,#0f172a 100%);border-radius:12px;padding:24px;border:1px solid #ffffff06}}
-    .chart-title{{font-size:.73rem;font-weight:700;letter-spacing:.07em;text-transform:uppercase;color:#64748b;margin-bottom:16px}}
-    .chart-wrap{{position:relative}}
-    .donut-wrap{{max-width:260px;margin:0 auto}}
-
-    /* GRAPH */
-    .graph-container{{position:relative;border-radius:12px;overflow:hidden;border:1px solid #ffffff08}}
-    .graph-container:fullscreen{{border-radius:0;background:#0f172a}}
-    .graph-container:fullscreen .cy-graph{{height:100vh}}
-    .cy-graph{{width:100%;height:600px;background:#0f172a}}
-    #cy{{width:100%;height:600px}}
-    #cyAttack{{width:100%;height:500px}}
-    .graph-controls{{position:absolute;top:12px;right:12px;display:flex;flex-direction:column;gap:4px;z-index:10}}
-    .graph-btn{{width:36px;height:36px;border-radius:8px;border:1px solid #334155;background:rgba(15,23,42,.85);color:#94a3b8;font-size:16px;cursor:pointer;display:flex;align-items:center;justify-content:center;transition:all .15s;backdrop-filter:blur(8px)}}
-    .graph-btn:hover{{background:#1e293b;color:#e2e8f0;border-color:#475569}}
-    .legend{{display:flex;gap:20px;flex-wrap:wrap;font-size:.76rem;color:#64748b;margin-top:12px;padding:0 4px}}
-    .legend span{{display:flex;align-items:center;gap:6px}}
-    .legend i{{display:inline-block;width:10px;height:10px;border-radius:3px}}
-    .legend i.diamond{{transform:rotate(45deg);border-radius:1px}}
-
-    /* MINIMAP */
-    .cy-minimap{{position:absolute;bottom:12px;left:12px;width:180px;height:130px;background:rgba(15,23,42,.9);border:1px solid #334155;border-radius:8px;overflow:hidden;z-index:10;backdrop-filter:blur(8px)}}
-    .cy-minimap canvas{{width:100%!important;height:100%!important}}
-
-    /* CONTEXT MENU */
-    .cy-ctx-menu{{position:fixed;background:rgba(15,23,42,.97);border:1px solid #334155;border-radius:10px;padding:6px 0;min-width:200px;z-index:300;backdrop-filter:blur(12px);box-shadow:0 8px 32px rgba(0,0,0,.5);display:none}}
-    .cy-ctx-menu.show{{display:block}}
-    .cy-ctx-item{{padding:8px 16px;font-size:.8rem;color:#cbd5e1;cursor:pointer;display:flex;align-items:center;gap:8px;transition:background .1s}}
-    .cy-ctx-item:hover{{background:#1e293b;color:#f1f5f9}}
-    .cy-ctx-sep{{height:1px;background:#1e293b;margin:4px 0}}
-
-    /* RISK PULSE for critical nodes */
-    @keyframes riskPulse{{0%{{box-shadow:0 0 0 0 rgba(220,38,38,.5)}}70%{{box-shadow:0 0 0 12px rgba(220,38,38,0)}}100%{{box-shadow:0 0 0 0 rgba(220,38,38,0)}}}}
-
-    /* NODE DETAIL SIDEBAR */
-    .node-sidebar{{position:fixed;top:0;right:0;bottom:0;width:340px;background:rgba(15,23,42,.97);border-left:1px solid #334155;backdrop-filter:blur(12px);z-index:200;overflow-y:auto;transform:translateX(100%);transition:transform .25s ease;padding:0}}
-    .node-sidebar.open{{transform:translateX(0);display:block}}
-    .sidebar-header{{display:flex;justify-content:space-between;align-items:center;padding:16px 20px 8px;border-bottom:1px solid #1e293b}}
-    .sidebar-type{{font-size:.65rem;letter-spacing:.08em;text-transform:uppercase;color:#64748b;font-weight:700;padding:3px 8px;border-radius:4px;border:1px solid #334155}}
-    .sidebar-close{{background:none;border:none;color:#64748b;font-size:1.4rem;cursor:pointer;padding:4px 8px;border-radius:4px;transition:all .15s}}
-    .sidebar-close:hover{{color:#e2e8f0;background:#1e293b}}
-    .sidebar-name{{font-size:1rem;font-weight:700;color:#f1f5f9;padding:12px 20px 4px;margin:0}}
-    .sidebar-meta{{font-size:.78rem;color:#94a3b8;padding:0 20px 12px;font-family:monospace;white-space:pre-line}}
-    .sidebar-section{{padding:0 20px 16px}}
-    .sidebar-section:empty{{display:none}}
-    .sidebar-label{{font-size:.68rem;letter-spacing:.06em;text-transform:uppercase;color:#64748b;font-weight:700;margin-bottom:8px}}
-    .sidebar-list{{list-style:none;padding:0;margin:0}}
-    .sidebar-list li{{font-size:.8rem;color:#cbd5e1;padding:5px 0;border-bottom:1px solid #0f172a}}
-    .sidebar-list li:last-child{{border-bottom:none}}
-    .sidebar-link{{color:#60a5fa;font-size:.78rem;text-decoration:none}}
-    .sidebar-link:hover{{text-decoration:underline;color:#93c5fd}}
-    .sidebar-cred{{color:#fbbf24;font-family:monospace;font-size:.78rem}}
-    @media(max-width:900px){{.node-sidebar{{width:100%}}}}
-
-    /* TOOLTIP */
-    #tip{{position:fixed;background:#0f172a;border:1px solid #334155;border-radius:8px;padding:10px 14px;font-size:.76rem;color:#e2e8f0;pointer-events:none;white-space:pre-line;max-width:280px;z-index:9999;display:none;line-height:1.5;box-shadow:0 8px 24px rgba(0,0,0,.4)}}
-
-    /* TABS */
-    .tab-bar{{display:flex;gap:2px;flex-wrap:wrap;margin:0 0 28px;border-bottom:1px solid #1e293b;position:sticky;top:0;background:#0b1120;z-index:60;padding-top:6px}}
-    .tab-btn{{background:transparent;border:none;border-bottom:2px solid transparent;color:#64748b;font-size:.82rem;font-weight:600;padding:11px 18px;cursor:pointer;transition:color .15s,border-color .15s;letter-spacing:.01em;white-space:nowrap}}
-    .tab-btn:hover{{color:#e2e8f0}}
-    .tab-btn.active{{color:#10b981;border-bottom-color:#10b981}}
-    .tab-count{{font-size:.66rem;background:#1e293b;color:#94a3b8;border-radius:10px;padding:1px 7px;margin-left:7px;font-weight:700}}
-    .tab-btn.active .tab-count{{background:rgba(16,185,129,.15);color:#6ee7b7}}
-    body.js-tabs .container>section[data-tab]{{display:none}}
-    body.js-tabs .container>section[data-tab].tab-active{{display:block}}
-
-    /* PAGINATION */
-    .pager{{display:flex;align-items:center;gap:8px;flex-wrap:wrap;margin-top:14px;padding:10px 14px;background:#0f172a;border:1px solid #1e293b;border-radius:8px;font-size:.76rem;color:#94a3b8}}
-    .pager-btn{{background:#1e293b;border:1px solid #334155;color:#cbd5e1;font-size:.74rem;font-weight:600;padding:5px 11px;border-radius:6px;cursor:pointer;transition:all .12s}}
-    .pager-btn:hover:not(:disabled){{background:#334155;color:#f1f5f9}}
-    .pager-btn:disabled{{opacity:.35;cursor:default}}
-    .pager-info{{color:#cbd5e1;font-variant-numeric:tabular-nums;padding:0 4px}}
-    .pager-sep{{flex:1}}
-    .pager-size-label{{display:flex;align-items:center;gap:6px;color:#64748b;text-transform:uppercase;letter-spacing:.05em;font-size:.68rem;font-weight:700}}
-    .pager-size{{background:#1e293b;border:1px solid #334155;border-radius:6px;color:#e2e8f0;font-size:.74rem;padding:4px 6px;outline:none}}
-    .pg-hidden{{display:none}}
-    @media print{{
-      .tab-bar,.pager{{display:none!important}}
-      body.js-tabs .container>section[data-tab]{{display:block!important}}
-      .pg-hidden{{display:table-row!important}}
-    }}
-
-    /* TABLES */
-    .table-wrap{{overflow-x:auto;border-radius:8px}}
-    .data-table{{width:100%;border-collapse:collapse;font-size:.83rem}}
-    .data-table th{{padding:10px 14px;font-size:.68rem;letter-spacing:.06em;color:#64748b;font-weight:700;text-transform:uppercase;border-bottom:2px solid #334155;white-space:nowrap;background:#0f172a;position:sticky;top:0}}
-    .data-table.sortable th{{cursor:pointer;user-select:none;transition:color .15s}}
-    .data-table.sortable th:hover{{color:#e2e8f0}}
-    .sort-arrow{{font-size:.6rem;margin-left:3px;opacity:.4}}
-    .sort-arrow.asc::after{{content:"\\25B2"}}
-    .sort-arrow.desc::after{{content:"\\25BC"}}
-    .data-table td{{padding:10px 14px;border-bottom:1px solid #1e293b;vertical-align:middle}}
-    .data-table tr{{transition:background .1s}}
-    .data-table tr:hover td{{background:rgba(255,255,255,.03)}}
-
-    /* BADGES */
-    .badge-kev{{background:#7f1d1d;color:#fca5a5;padding:2px 8px;border-radius:4px;font-size:.68rem;font-weight:700}}
-    .badge-exploit-likely{{background:#7c2d12;color:#fdba74;padding:2px 8px;border-radius:4px;font-size:.68rem;font-weight:700}}
-    .badge-exploit-public{{background:#713f12;color:#fde047;padding:2px 8px;border-radius:4px;font-size:.66rem;font-weight:600}}
-    .badge-ai{{background:#1d4ed8;color:#bfdbfe;padding:2px 8px;border-radius:4px;font-size:.68rem;font-weight:700;margin-right:4px}}
-    .badge-vuln{{background:#7f1d1d;color:#fca5a5;font-size:.65rem;padding:2px 6px;border-radius:4px;font-weight:700}}
-    .badge-cred{{background:#78350f;color:#fde68a;font-size:.65rem;padding:2px 6px;border-radius:4px;font-weight:700}}
-    .vuln-id{{color:#93c5fd;font-size:.78rem}}
-
-    /* REMEDIATION */
-    .exposure-paths{{display:grid;gap:12px}}
-    .exposure-path-card{{display:flex;gap:14px;padding:14px 0;border-bottom:1px solid #1e293b}}
-    .exposure-path-card:last-child{{border-bottom:none}}
-    .path-rank{{width:42px;height:42px;border:2px solid #334155;border-radius:8px;display:flex;align-items:center;justify-content:center;font-weight:800;flex-shrink:0;background:#0f172a}}
-    .path-body{{min-width:0;flex:1}}
-    .path-title{{color:#e2e8f0;font-weight:700;font-size:.92rem;margin-bottom:4px}}
-    .path-summary{{color:#94a3b8;font-size:.82rem;line-height:1.45;margin-bottom:8px}}
-    .path-meta{{display:flex;flex-wrap:wrap;gap:8px;color:#64748b;font-size:.74rem;margin-bottom:8px}}
-    .path-meta span{{background:#0f172a;border:1px solid #1e293b;border-radius:999px;padding:3px 8px}}
-    .path-fix{{color:#4ade80;font-size:.78rem;font-weight:600}}
-    .remediation-item{{display:flex;align-items:flex-start;gap:14px;padding:16px 0;border-bottom:1px solid #1e293b;transition:background .1s}}
-    .remediation-item:hover{{background:rgba(255,255,255,.02);margin:0 -12px;padding-left:12px;padding-right:12px;border-radius:8px}}
-    .subsection-label{{font-size:.7rem;letter-spacing:.07em;text-transform:uppercase;color:#64748b;margin-bottom:10px}}
-
-    /* INVENTORY */
-    .inv-search{{width:100%;padding:10px 14px;background:#0f172a;border:1px solid #334155;border-radius:8px;color:#e2e8f0;font-size:.85rem;margin-bottom:16px;outline:none;transition:border-color .15s}}
-    .inv-search:focus{{border-color:#3b82f6}}
-    .inv-search::placeholder{{color:#475569}}
-    .agent-card{{background:linear-gradient(135deg,#1e293b 0%,#0f172a 100%);border-radius:12px;margin-bottom:12px;overflow:hidden;border:1px solid #ffffff06;transition:box-shadow .15s}}
-    .agent-card:hover{{box-shadow:0 4px 16px rgba(0,0,0,.2)}}
-    .agent-summary{{list-style:none;display:flex;justify-content:space-between;align-items:center;padding:18px 22px;cursor:pointer;user-select:none;font-weight:700;font-size:.95rem;color:#f1f5f9}}
-    .agent-summary::-webkit-details-marker{{display:none}}
-    .agent-summary::before{{content:"\\25B6";margin-right:10px;font-size:.6rem;color:#475569;transition:transform .2s}}
-    details[open] .agent-summary::before{{transform:rotate(90deg)}}
-    .agent-detail{{padding:18px 22px;border-top:1px solid #0b112060}}
-    .server-card{{background:#0b1120;border-radius:8px;padding:14px 16px;margin-bottom:10px;border-left:3px solid #334155;border:1px solid #ffffff04;border-left:3px solid #334155}}
-    .pkg-row{{display:flex;justify-content:space-between;padding:5px 0;border-bottom:1px solid #0a162830;font-size:.78rem}}
-    .pkg-row:last-child{{border-bottom:none}}
-    .pkg-name{{color:#e2e8f0}}
-    .pkg-ver{{color:#64748b;font-family:monospace;font-size:.73rem}}
-    .toggle-btn{{background:transparent;border:1px solid #334155;color:#64748b;font-size:.72rem;padding:6px 12px;border-radius:6px;cursor:pointer;margin-top:10px;width:100%;transition:all .15s}}
-    .toggle-btn:hover{{background:#1e293b;color:#94a3b8}}
-
-    /* HINTS */
-    .hint-box{{background:#1e3a5f40;border:1px solid #3b82f640;border-radius:8px;padding:14px 18px;margin-bottom:18px;font-size:.82rem;color:#93c5fd}}
-    .empty-state{{background:#052e1615;border:1px solid #16a34a30;border-radius:10px;padding:24px;color:#4ade80;text-align:center;font-size:.9rem}}
-
-    footer{{border-top:1px solid #1e293b;padding:24px 32px;text-align:center;font-size:.75rem;color:#334155;margin-left:220px}}
-    .print-btn{{background:transparent;border:1px solid #334155;color:#64748b;font-size:.75rem;padding:4px 12px;border-radius:6px;cursor:pointer;margin-left:12px;transition:all .15s}}
-    .print-btn:hover{{background:#1e293b;color:#94a3b8}}
-
-    @media(max-width:900px){{
-      .sidebar{{display:none}}
-      .sidebar.mobile-open{{display:flex}}
-      .sidebar-toggle{{display:flex;align-items:center;justify-content:center}}
-      .container{{margin-left:0;padding:60px 16px 60px}}
-      footer{{margin-left:0}}
-      .charts-row{{grid-template-columns:1fr}}
-      .stat-grid{{grid-template-columns:repeat(auto-fill,minmax(140px,1fr))}}
-      .node-sidebar{{width:100%}}
-    }}
-
-    /* PRINT */
-    @media print{{
-      body{{background:#fff;color:#1e293b;font-size:12px}}
-      .sidebar,.graph-controls,.graph-filter-bar,.vuln-filter-bar,.policy-filter-bar,.toggle-btn,.inv-search,.print-btn,.node-sidebar,.sidebar-toggle{{display:none!important}}
-      .container{{margin-left:0;max-width:100%;padding:10px}}
-      footer{{margin-left:0}}
-      section{{page-break-inside:avoid;margin-bottom:20px}}
-      .panel,.stat-card,.agent-card,.server-card,.chart-panel{{background:#f8fafc;border:1px solid #e2e8f0;box-shadow:none}}
-      .stat-value,.sec-title{{color:#0f172a}}
-      .data-table th{{background:#f1f5f9;color:#334155;border-bottom:2px solid #cbd5e1}}
-      .data-table td{{border-bottom:1px solid #e2e8f0}}
-      .path-rank,.path-meta span{{background:#f1f5f9;border-color:#e2e8f0}}
-      .path-title{{color:#0f172a}}
-      .path-summary{{color:#475569}}
-      .path-fix{{color:#15803d}}
-      #cy{{height:400px;background:#f8fafc;border:1px solid #e2e8f0}}
-      .legend{{color:#475569}}
-      a{{color:#2563eb}}
-      footer{{color:#94a3b8}}
-      .graph-container{{border:1px solid #e2e8f0}}
-    }}
-  </style>
-</head>
-<body>
-
-<!-- Mobile sidebar toggle -->
-<button class="sidebar-toggle" id="sidebarToggle" onclick="document.getElementById('mainSidebar').classList.toggle('mobile-open');this.innerHTML=this.innerHTML==='&#9776;'?'&times;':'&#9776;'">&#9776;</button>
-
-<!-- Sidebar Navigation -->
-<div class="sidebar" id="mainSidebar">
-  <div class="sidebar-brand">
-    <div class="brand-icon">&#x1f6e1;&#xfe0f;</div>
-    <div>
-      <div class="brand-text">agent-bom</div>
-      <div class="brand-sub">AI Supply Chain Security</div>
-    </div>
-  </div>
-  <div class="sidebar-status">
-    <span class="status-badge">{status_label}</span>
-    <span class="scan-time">v{_esc(report.tool_version)}</span>
-  </div>
-
-  <div class="sidebar-group">
-    <div class="sidebar-group-label">Overview</div>
-    <a href="#summary" class="sidebar-link active"><span class="link-icon">&#x1f4ca;</span> Summary</a>
-    <a href="#charts" class="sidebar-link"><span class="link-icon">&#x1f4c8;</span> Risk Charts</a>
-  </div>
-
-  <div class="sidebar-group">
-    <div class="sidebar-group-label">Analysis</div>
-    <a href="#riskmap" class="sidebar-link"><span class="link-icon">&#x1f5fa;&#xfe0f;</span> Supply Chain Graph</a>
-    <a href="#inventory" class="sidebar-link"><span class="link-icon">&#x1f4e6;</span> Agent Inventory</a>
-    {'<a href="#aiinventory" class="sidebar-link"><span class="link-icon">&#x1f916;</span> AI Inventory</a>' if ai_inv_section else ""}
-  </div>
-
-  <div class="sidebar-group">
-    <div class="sidebar-group-label">Security</div>
-    {'<a href="#attackflow" class="sidebar-link"><span class="link-icon">&#x26a1;</span> Attack Flow</a>' if findings else ""}
-    {f'<a href="#vulns" class="sidebar-link"><span class="link-icon">&#x1f41b;</span> Vulnerabilities <span class="link-badge" style="background:#7f1d1d;color:#fca5a5">{len(findings)}</span></a>' if findings else ""}
-    {f'<a href="#policyfindings" class="sidebar-link"><span class="link-icon">&#x1f6e1;&#xfe0f;</span> Policy Findings <span class="link-badge" style="background:#78350f;color:#fcd34d">{len(policy_findings)}</span></a>' if policy_findings else ""}
-    {'<a href="#exposure-paths" class="sidebar-link"><span class="link-icon">&#x1f9ed;</span> Exposure Paths</a>' if findings else ""}
-    {'<a href="#blast" class="sidebar-link"><span class="link-icon">&#x1f4a5;</span> Blast Radius</a>' if findings else ""}
-    {'<a href="#remediation" class="sidebar-link"><span class="link-icon">&#x1f527;</span> Remediation</a>' if findings else ""}
-  </div>
-
-  <div class="sidebar-group">
-    <div class="sidebar-group-label">Governance</div>
-    {'<a href="#compliance" class="sidebar-link"><span class="link-icon">&#x2705;</span> Compliance</a>' if compliance_html else ""}
-    {'<a href="#skillaudit" class="sidebar-link"><span class="link-icon">&#x1f50d;</span> Skill Audit</a>' if skill_section else ""}
-    {'<a href="#trust" class="sidebar-link"><span class="link-icon">&#x1f91d;</span> Trust</a>' if trust_section else ""}
-    {'<a href="#enforcement" class="sidebar-link"><span class="link-icon">&#x1f512;</span> Enforcement</a>' if enforce_section else ""}
-    {'<a href="#cisbenchmarks" class="sidebar-link"><span class="link-icon">&#x1f6e1;&#xfe0f;</span> CIS Benchmarks</a>' if cis_bench_section else ""}
-  </div>
-
-  <div class="sidebar-spacer"></div>
-
-  <div class="sidebar-footer">
-    <div class="scan-time">{_esc(generated)}</div>
-    <button class="print-btn" onclick="window.print()">&#x1f5b6;&#xfe0f; Print / Export PDF</button>
-  </div>
-</div>
-
-<div id="tip"></div>
-
-<div id="nodeDetailSidebar" class="node-sidebar" style="display:none">
-  <div class="sidebar-header">
-    <span id="sidebarNodeType" class="sidebar-type"></span>
-    <button id="sidebarClose" class="sidebar-close">&times;</button>
-  </div>
-  <h3 id="sidebarNodeName" class="sidebar-name"></h3>
-  <div id="sidebarMeta" class="sidebar-meta"></div>
-  <div id="sidebarConnected" class="sidebar-section"></div>
-  <div id="sidebarCredentials" class="sidebar-section"></div>
-  <div id="sidebarCves" class="sidebar-section"></div>
-  <div id="sidebarRemediation" class="sidebar-section"></div>
-</div>
-
-<div class="container">
-
-  <!-- Delta / warn-gate banners (only rendered when applicable) -->
-  {_delta_banner(report)}
-  {_warn_gate_banner(report)}
-
-  <!-- Summary stat cards -->
-  <section id="summary">
-    <div class="sec-title">Summary</div>
-    {_summary_cards(report, findings, policy_findings)}
-  </section>
-
-  <!-- Charts row -->
-  <section id="charts">
-    <div class="sec-title">Risk Overview</div>
-    <div class="charts-row">
-      <div class="chart-panel">
-        <div class="chart-title">Severity Distribution</div>
-        <div class="donut-wrap">
-          <canvas id="sevChart" height="240"></canvas>
-        </div>
-      </div>
-      <div class="chart-panel">
-        <div class="chart-title">Top Blast Radius Scores</div>
-        <div class="chart-wrap">
-          <canvas id="blastChart" height="240"></canvas>
-        </div>
-      </div>
-    </div>
-  </section>
-
-  <!-- Risk map graph -->
-  <section id="riskmap">
-    <div class="sec-title">
-      Supply Chain Graph
-      <span style="font-size:.68rem;font-weight:400;opacity:.5;margin-left:8px">
-        {_esc(graph_note)}
-      </span>
-    </div>
-    <div class="graph-container">
-      <div id="cy"></div>
-      <div class="graph-filter-bar" style="position:absolute;top:12px;left:12px;display:flex;gap:8px;align-items:center;z-index:10;background:rgba(15,23,42,.92);padding:8px 12px;border-radius:8px;border:1px solid #334155;backdrop-filter:blur(8px)">
-        <label style="display:flex;align-items:center;gap:3px;font-size:.72rem;color:#fca5a5;cursor:pointer"><input type="checkbox" class="graph-sev-filter" value="critical" checked> Crit</label>
-        <label style="display:flex;align-items:center;gap:3px;font-size:.72rem;color:#fb923c;cursor:pointer"><input type="checkbox" class="graph-sev-filter" value="high" checked> High</label>
-        <label style="display:flex;align-items:center;gap:3px;font-size:.72rem;color:#fbbf24;cursor:pointer"><input type="checkbox" class="graph-sev-filter" value="medium" checked> Med</label>
-        <label style="display:flex;align-items:center;gap:3px;font-size:.72rem;color:#94a3b8;cursor:pointer"><input type="checkbox" class="graph-sev-filter" value="low" checked> Low</label>
-        <span style="width:1px;height:16px;background:#334155"></span>
-        <input type="text" id="graphSearch" placeholder="Search nodes&hellip;" style="padding:4px 8px;background:#0f172a;border:1px solid #334155;border-radius:4px;color:#e2e8f0;font-size:.72rem;width:120px;outline:none">
-      </div>
-      <div class="graph-controls">
-        <button class="graph-btn" id="zoomIn" title="Zoom in">+</button>
-        <button class="graph-btn" id="zoomOut" title="Zoom out">&minus;</button>
-        <button class="graph-btn" id="fitBtn" title="Fit to view">&#x2922;</button>
-        <button class="graph-btn" id="fullscreenBtn" title="Fullscreen">&#x26F6;</button>
-      </div>
-    </div>
-    <div class="legend">
-      <span><i style="background:#818cf8"></i>Provider</span>
-      <span><i style="background:#3b82f6"></i>Agent</span>
-      <span><i style="background:#10b981"></i>Server (clean)</span>
-      <span><i style="background:#f59e0b"></i>Server (credentials)</span>
-      <span><i style="background:#ef4444"></i>Server (vulnerable)</span>
-      <span><i style="background:#dc2626"></i>Vulnerable package</span>
-      <span><i class="diamond" style="background:#f87171"></i>CVE</span>
-    </div>
-  </section>
-
-  <!-- Agent inventory (collapsible) -->
-  <section id="inventory">
-    <div class="sec-title">Agent Inventory</div>
-    <input type="text" class="inv-search" id="invSearch" placeholder="Search agents, servers, packages&hellip;">
-    {_inventory_cards(report)}
-  </section>
-
-  <!-- AI component inventory -->
-  {ai_inv_section}
-
-  <!-- Skill audit -->
-  {skill_section}
-
-  <!-- Trust assessment -->
-  {trust_section}
-
-  <!-- Enforcement -->
-  {enforce_section}
-
-  <!-- CIS benchmark posture (issue #665) -->
-  {cis_bench_section}
-
-  <!-- Attack flow graph (only when vulns exist) -->
-  {_attack_flow_section(findings)}
-
-  <!-- Vuln / Blast / Remediation -->
-  {policy_section}
-  {vuln_sections}
-
-  <!-- Compliance posture -->
-  {compliance_html}
-
-</div>
-
-<footer>
-  Generated by <strong style="color:#475569">agent-bom</strong> v{_esc(report.tool_version)} &middot;
-  <a href="https://github.com/msaad00/agent-bom">github.com/msaad00/agent-bom</a> &middot;
-  Vulnerability data: OSV.dev &middot; NVD &middot; CISA KEV &middot; EPSS
-</footer>
-
-<script>
-// agent-bom scale-report: tabs + client-side pagination. Kept in a standalone
-// script (distinct opening so offline mode does not strip it) and independent of
-// the CDN chart/graph libs, so a large report stays tabbed and paginated even
-// when opened offline or from an email attachment.
-(function scaleReport() {{
-  window.PAGINATORS = window.PAGINATORS || {{}};
-
-  function makePaginator(tableId, filterFn) {{
-    var table = document.getElementById(tableId);
-    if (!table) return null;
-    var tbody = table.querySelector('tbody');
-    var bar = document.querySelector('.pager[data-pager="' + tableId + '"]');
-    var allRows = Array.prototype.slice.call(tbody.querySelectorAll('tr'));
-    var pageSize = bar ? (parseInt(bar.getAttribute('data-page-size'), 10) || 50) : 50;
-    var page = 1;
-    var matched = allRows;
-    var infoEl = bar ? bar.querySelector('.pager-info') : null;
-    function render() {{
-      var total = matched.length;
-      var pages = Math.max(1, Math.ceil(total / pageSize));
-      if (page > pages) page = pages;
-      if (page < 1) page = 1;
-      var start = (page - 1) * pageSize;
-      var end = start + pageSize;
-      allRows.forEach(function(r) {{ r.classList.add('pg-hidden'); }});
-      matched.slice(start, end).forEach(function(r) {{ r.classList.remove('pg-hidden'); }});
-      if (bar) {{
-        if (infoEl) infoEl.innerHTML = total ? (start + 1) + '&ndash;' + Math.min(end, total) + ' of ' + total : '0 of 0';
-        var f = bar.querySelector('[data-act="first"]'), pv = bar.querySelector('[data-act="prev"]'),
-            nx = bar.querySelector('[data-act="next"]'), ls = bar.querySelector('[data-act="last"]');
-        if (f) f.disabled = page <= 1;
-        if (pv) pv.disabled = page <= 1;
-        if (nx) nx.disabled = page >= pages;
-        if (ls) ls.disabled = page >= pages;
-        bar.style.display = pages > 1 ? '' : 'none';
-      }}
-    }}
-    function apply() {{ matched = filterFn ? allRows.filter(filterFn) : allRows; page = 1; render(); }}
-    function resort() {{ allRows = Array.prototype.slice.call(tbody.querySelectorAll('tr')); apply(); }}
-    if (bar) {{
-      bar.addEventListener('click', function(e) {{
-        var act = e.target.getAttribute && e.target.getAttribute('data-act');
-        if (!act) return;
-        var pages = Math.max(1, Math.ceil(matched.length / pageSize));
-        if (act === 'first') page = 1;
-        else if (act === 'prev') page = Math.max(1, page - 1);
-        else if (act === 'next') page = Math.min(pages, page + 1);
-        else if (act === 'last') page = pages;
-        render();
-      }});
-      var sizeSel = bar.querySelector('.pager-size');
-      if (sizeSel) sizeSel.addEventListener('change', function() {{ pageSize = parseInt(this.value, 10) || 50; page = 1; render(); }});
-    }}
-    var api = {{ apply: apply, resort: resort, render: render }};
-    window.PAGINATORS[tableId] = api;
-    apply();
-    return api;
-  }}
-
-  // Vulnerability table filter (drives its paginator).
-  function vulnRowMatch(row) {{
-    var checkedSevs = Array.prototype.slice.call(document.querySelectorAll('.vuln-sev-filter:checked')).map(function(c) {{ return c.value; }});
-    var kevOnly = document.getElementById('kevToggle') && document.getElementById('kevToggle').checked;
-    var q = ((document.getElementById('vulnSearch') || {{}}).value || '').toLowerCase();
-    var sev = row.getAttribute('data-severity') || '';
-    if (checkedSevs.indexOf(sev) === -1) return false;
-    if (kevOnly && row.getAttribute('data-kev') !== '1') return false;
-    if (q && row.textContent.toLowerCase().indexOf(q) === -1) return false;
-    return true;
-  }}
-  function filterVulnTable() {{ if (window.PAGINATORS.vulnTable) window.PAGINATORS.vulnTable.apply(); }}
-  makePaginator('vulnTable', vulnRowMatch);
-  document.querySelectorAll('.vuln-sev-filter').forEach(function(cb) {{ cb.addEventListener('change', filterVulnTable); }});
-  var kevToggle = document.getElementById('kevToggle');
-  if (kevToggle) kevToggle.addEventListener('change', filterVulnTable);
-  var vulnSearchInput = document.getElementById('vulnSearch');
-  if (vulnSearchInput) vulnSearchInput.addEventListener('input', filterVulnTable);
-
-  // Unified policy/security finding filter (drives its paginator).
-  function policyRowMatch(row) {{
-    var checkedSevs = Array.prototype.slice.call(document.querySelectorAll('.policy-sev-filter:checked')).map(function(c) {{ return c.value; }});
-    var typeFilter = (document.getElementById('policyTypeFilter') || {{}}).value || '';
-    var assetFilter = (document.getElementById('policyAssetFilter') || {{}}).value || '';
-    var q = ((document.getElementById('policySearch') || {{}}).value || '').toLowerCase();
-    var sev = row.getAttribute('data-severity') || '';
-    if (checkedSevs.indexOf(sev) === -1) return false;
-    if (typeFilter && (row.getAttribute('data-type') || '') !== typeFilter) return false;
-    if (assetFilter && (row.getAttribute('data-asset-type') || '') !== assetFilter) return false;
-    if (q && row.textContent.toLowerCase().indexOf(q) === -1) return false;
-    return true;
-  }}
-  function filterPolicyFindingsTable() {{
-    if (window.PAGINATORS.policyFindingsTable) window.PAGINATORS.policyFindingsTable.apply();
-    var count = document.getElementById('policyVisibleCount');
-    var table = document.getElementById('policyFindingsTable');
-    if (count && table) {{
-      var all = table.querySelectorAll('tbody tr');
-      var vis = Array.prototype.slice.call(all).filter(policyRowMatch).length;
-      count.textContent = vis + ' of ' + all.length + ' shown';
-    }}
-  }}
-  makePaginator('policyFindingsTable', policyRowMatch);
-  document.querySelectorAll('.policy-sev-filter').forEach(function(cb) {{ cb.addEventListener('change', filterPolicyFindingsTable); }});
-  var policyTypeFilter = document.getElementById('policyTypeFilter');
-  if (policyTypeFilter) policyTypeFilter.addEventListener('change', filterPolicyFindingsTable);
-  var policyAssetFilter = document.getElementById('policyAssetFilter');
-  if (policyAssetFilter) policyAssetFilter.addEventListener('change', filterPolicyFindingsTable);
-  var policySearchInput = document.getElementById('policySearch');
-  if (policySearchInput) policySearchInput.addEventListener('input', filterPolicyFindingsTable);
-  filterPolicyFindingsTable();
-
-  // ── Tabbed navigation ─────────────────────────────────────────────────────
-  var tabBar = document.querySelector('.tab-bar');
-  function tabForSection(id) {{
-    var sec = document.getElementById(id);
-    return sec ? sec.getAttribute('data-tab') : null;
-  }}
-  function activateTab(key) {{
-    if (!key) return;
-    document.querySelectorAll('.tab-btn').forEach(function(b) {{ b.classList.toggle('active', b.getAttribute('data-tab') === key); }});
-    document.querySelectorAll('.container>section[data-tab]').forEach(function(s) {{ s.classList.toggle('tab-active', s.getAttribute('data-tab') === key); }});
-    // Tables re-render because a hidden tab had zero layout width.
-    Object.keys(window.PAGINATORS).forEach(function(id) {{ window.PAGINATORS[id].render(); }});
-  }}
-  if (tabBar) {{
-    document.body.classList.add('js-tabs');
-    tabBar.querySelectorAll('.tab-btn').forEach(function(b) {{
-      b.addEventListener('click', function() {{ activateTab(b.getAttribute('data-tab')); window.scrollTo(0, 0); }});
-    }});
-    var firstTab = tabBar.querySelector('.tab-btn');
-    if (firstTab) activateTab(firstTab.getAttribute('data-tab'));
-  }}
-
-  // Sidebar / in-page anchors reveal the target's tab, then scroll to it.
-  document.querySelectorAll('a[href^="#"]').forEach(function(a) {{
-    a.addEventListener('click', function(e) {{
-      var id = a.getAttribute('href').slice(1);
-      if (!id) return;
-      var el = document.getElementById(id);
-      var key = tabForSection(id);
-      if (tabBar && key) {{
-        e.preventDefault();
-        activateTab(key);
-        if (el) setTimeout(function() {{ el.scrollIntoView({{ behavior: 'smooth', block: 'start' }}); }}, 30);
-      }}
-    }});
-  }});
-}})();
-</script>
-
-<script>
-(function() {{
-  // Injected data
-  var CHART_DATA = {chart_data_json};
-  var GRAPH_ELEMENTS = {elements_json};
-  var ATTACK_FLOW = {attack_flow_json};
-
-  // Chart.js: Severity donut
-  var sevCtx = document.getElementById('sevChart');
-  if (sevCtx && CHART_DATA.sev.data.some(function(v){{ return v > 0; }})) {{
-    new Chart(sevCtx, {{
-      type: 'doughnut',
-      data: {{
-        labels: CHART_DATA.sev.labels,
-        datasets: [{{
-          data: CHART_DATA.sev.data,
-          backgroundColor: CHART_DATA.sev.colors,
-          borderColor: '#0b1120',
-          borderWidth: 3,
-          hoverOffset: 8,
-        }}],
-      }},
-      options: {{
-        responsive: true,
-        cutout: '68%',
-        plugins: {{
-          legend: {{
-            position: 'bottom',
-            labels: {{
-              color: '#94a3b8',
-              font: {{ size: 11 }},
-              boxWidth: 12,
-              padding: 14,
-            }},
-          }},
-          tooltip: {{
-            backgroundColor: '#0f172a',
-            borderColor: '#334155',
-            borderWidth: 1,
-            titleColor: '#e2e8f0',
-            bodyColor: '#94a3b8',
-            cornerRadius: 8,
-            padding: 10,
-            callbacks: {{
-              label: function(ctx) {{
-                return ' ' + ctx.label + ': ' + ctx.parsed;
-              }},
-            }},
-          }},
-        }},
-      }},
-    }});
-  }} else if (sevCtx) {{
-    var p = document.createElement('p');
-    p.style.cssText = 'color:#4ade80;text-align:center;padding:50px 0;font-size:.88rem';
-    p.innerHTML = '&#x2705; No vulnerabilities';
-    sevCtx.parentNode.replaceChild(p, sevCtx);
-  }}
-
-  // Chart.js: Blast radius bar
-  var blastCtx = document.getElementById('blastChart');
-  if (blastCtx && CHART_DATA.blast.labels.length > 0) {{
-    new Chart(blastCtx, {{
-      type: 'bar',
-      data: {{
-        labels: CHART_DATA.blast.labels,
-        datasets: [{{
-          label: 'Blast Score',
-          data: CHART_DATA.blast.scores,
-          backgroundColor: CHART_DATA.blast.colors,
-          borderRadius: 6,
-          borderSkipped: false,
-        }}],
-      }},
-      options: {{
-        indexAxis: 'y',
-        responsive: true,
-        scales: {{
-          x: {{
-            min: 0, max: 10,
-            grid: {{ color: '#1e293b' }},
-            ticks: {{ color: '#64748b', font: {{ size: 11 }} }},
-          }},
-          y: {{
-            grid: {{ display: false }},
-            ticks: {{ color: '#94a3b8', font: {{ size: 11 }} }},
-          }},
-        }},
-        plugins: {{
-          legend: {{ display: false }},
-          tooltip: {{
-            backgroundColor: '#0f172a',
-            borderColor: '#334155',
-            borderWidth: 1,
-            titleColor: '#e2e8f0',
-            bodyColor: '#94a3b8',
-            cornerRadius: 8,
-            callbacks: {{
-              label: function(ctx) {{
-                return ' Score: ' + ctx.parsed.x.toFixed(2);
-              }},
-            }},
-          }},
-        }},
-      }},
-    }});
-  }} else if (blastCtx) {{
-    var p2 = document.createElement('p');
-    p2.style.cssText = 'color:#4ade80;text-align:center;padding:50px 0;font-size:.88rem';
-    p2.innerHTML = '&#x2705; No blast radius data';
-    blastCtx.parentNode.replaceChild(p2, blastCtx);
-  }}
-
-  // Cytoscape: Supply chain graph with dagre hierarchical layout
-  var cyContainer = document.getElementById('cy');
-  if (cyContainer && GRAPH_ELEMENTS.length > 0) {{
-    var cy = cytoscape({{
-      container: cyContainer,
-      elements: GRAPH_ELEMENTS,
-      style: [
-        {{
-          selector: 'node[type="provider"]',
-          style: {{
-            'background-color': '#1e1b4b',
-            'border-color': '#818cf8',
-            'border-width': 3,
-            'label': 'data(label)',
-            'color': '#c7d2fe',
-            'font-size': '13px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 140,
-            'height': 44,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '125px',
-          }},
-        }},
-        {{
-          selector: 'node[type="agent"]',
-          style: {{
-            'background-color': '#1e3a8a',
-            'border-color': '#3b82f6',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#bfdbfe',
-            'font-size': '12px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 40,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '105px',
-          }},
-        }},
-        {{
-          selector: 'node[type="server_clean"]',
-          style: {{
-            'background-color': '#052e16',
-            'border-color': '#10b981',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#6ee7b7',
-            'font-size': '10px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 36,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '110px',
-          }},
-        }},
-        {{
-          selector: 'node[type="server_cred"]',
-          style: {{
-            'background-color': '#431407',
-            'border-color': '#f59e0b',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fde68a',
-            'font-size': '10px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 36,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '110px',
-          }},
-        }},
-        {{
-          selector: 'node[type="server_vuln"]',
-          style: {{
-            'background-color': '#450a0a',
-            'border-color': '#ef4444',
-            'border-width': 2.5,
-            'label': 'data(label)',
-            'color': '#fca5a5',
-            'font-size': '10px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 36,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '110px',
-          }},
-        }},
-        {{
-          selector: 'node[type="pkg_vuln"]',
-          style: {{
-            'background-color': '#7f1d1d',
-            'border-color': '#dc2626',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fca5a5',
-            'font-size': '9px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 130,
-            'height': 38,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '120px',
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="critical"], node[type^="cve_critical"]',
-          style: {{
-            'background-color': '#991b1b',
-            'border-color': '#f87171',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fecaca',
-            'font-size': '8px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 110,
-            'height': 30,
-            'shape': 'diamond',
-            'underlay-color': '#ef4444',
-            'underlay-padding': '6px',
-            'underlay-opacity': 0.15,
-            'underlay-shape': 'ellipse',
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="high"], node[type^="cve_high"]',
-          style: {{
-            'background-color': '#9a3412',
-            'border-color': '#fb923c',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fed7aa',
-            'font-size': '8px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 100,
-            'height': 28,
-            'shape': 'diamond',
-            'underlay-color': '#fb923c',
-            'underlay-padding': '4px',
-            'underlay-opacity': 0.1,
-            'underlay-shape': 'ellipse',
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="medium"], node[type="cve"][severity="low"], node[type="cve"][severity="none"], node[type^="cve_medium"], node[type^="cve_low"], node[type^="cve_none"]',
-          style: {{
-            'background-color': '#854d0e',
-            'border-color': '#fbbf24',
-            'border-width': 1.5,
-            'label': 'data(label)',
-            'color': '#fef08a',
-            'font-size': '8px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 90,
-            'height': 26,
-            'shape': 'diamond',
-          }},
-        }},
-        {{
-          selector: 'edge',
-          style: {{
-            'width': 1.8,
-            'line-color': '#334155',
-            'target-arrow-color': '#475569',
-            'target-arrow-shape': 'triangle',
-            'curve-style': 'bezier',
-            'arrow-scale': 0.8,
-          }},
-        }},
-        {{
-          selector: 'edge[type="hosts"]',
-          style: {{
-            'line-color': '#818cf850',
-            'target-arrow-color': '#818cf880',
-            'line-style': 'dashed',
-            'line-dash-pattern': [6, 3],
-          }},
-        }},
-        {{
-          selector: 'edge[type="affects"]',
-          style: {{
-            'line-color': '#dc262650',
-            'target-arrow-color': '#dc262680',
-          }},
-        }},
-        {{
-          selector: '.highlighted',
-          style: {{
-            'border-width': 4,
-            'border-color': '#f1f5f9',
-            'z-index': 999,
-          }},
-        }},
-        {{
-          selector: '.faded',
-          style: {{ 'opacity': 0.08 }},
-        }},
-      ],
-      layout: {{
-        name: 'dagre',
-        rankDir: 'LR',
-        nodeSep: 50,
-        rankSep: 80,
-        edgeSep: 15,
-        padding: 30,
-        animate: false,
-        fit: true,
-      }},
-      minZoom: 0.15,
-      maxZoom: 4,
-      wheelSensitivity: 0.3,
-      autoungrabify: false,
-    }});
-    cy.ready(function() {{ cy.fit(cy.elements(), 40); }});
-
-    // Tooltip
-    var tip = document.getElementById('tip');
-    cy.on('mouseover', 'node', function(e) {{
-      var t = e.target.data('tip');
-      if (t) {{ tip.textContent = t; tip.style.display = 'block'; }}
-    }});
-    cy.on('mousemove', function(e) {{
-      if (tip.style.display === 'block') {{
-        tip.style.left = (e.originalEvent.clientX + 14) + 'px';
-        tip.style.top  = (e.originalEvent.clientY + 14) + 'px';
-      }}
-    }});
-    cy.on('mouseout', 'node', function() {{ tip.style.display = 'none'; }});
-
-    // Click to highlight + sidebar
-    var sidebar = document.getElementById('nodeDetailSidebar');
-    var sidebarCloseBtn = document.getElementById('sidebarClose');
-
-    function escHtml(value) {{
-      return String(value == null ? '' : value).replace(/[&<>"']/g, function(ch) {{
-        return {{'&':'&amp;','<':'&lt;','>':'&gt;','"':'&quot;',"'":'&#39;'}}[ch];
-      }});
-    }}
-
-    function urlPart(value) {{
-      return encodeURIComponent(String(value == null ? '' : value));
-    }}
-
-    function showSidebar(node) {{
-      var d = node.data();
-      var t = d.type || '';
-      var isCveNode = t === 'cve' || t.indexOf('cve_')===0;
-      var typeLabels = {{'provider':'Provider','agent':'Agent','server_clean':'MCP Server','server_cred':'MCP Server','server_vuln':'MCP Server','pkg_vuln':'Package','cve':'Vulnerability'}};
-      var typeLabel = typeLabels[t] || (isCveNode ? 'Vulnerability' : t);
-      var typeColors = {{'provider':'#818cf8','agent':'#3b82f6','server_clean':'#10b981','server_cred':'#f59e0b','server_vuln':'#ef4444','pkg_vuln':'#dc2626'}};
-      var badgeColor = typeColors[t] || (isCveNode ? '#f87171' : '#64748b');
-
-      document.getElementById('sidebarNodeType').textContent = typeLabel;
-      document.getElementById('sidebarNodeType').style.borderColor = badgeColor;
-      document.getElementById('sidebarNodeType').style.color = badgeColor;
-      document.getElementById('sidebarNodeName').textContent = d.label || d.id;
-
-      ['sidebarMeta','sidebarConnected','sidebarCredentials','sidebarCves','sidebarRemediation'].forEach(function(id) {{
-        document.getElementById(id).innerHTML = '';
-      }});
-
-      // Connected nodes
-      var neighbors = node.neighborhood('node');
-      if (neighbors.length > 0) {{
-        var h = '<div class="sidebar-label">Connected (' + neighbors.length + ')</div><ul class="sidebar-list">';
-        neighbors.forEach(function(n) {{
-          var nt = n.data('type') || '';
-          var icon = nt === 'agent' ? '&#x1f916;' : nt.indexOf('server')===0 ? '&#x2699;' : nt === 'pkg_vuln' ? '&#x1f4e6;' : (nt === 'cve' || nt.indexOf('cve_')===0) ? '&#x1f41b;' : '&#x25cf;';
-          h += '<li>' + icon + ' ' + escHtml((n.data('label') || n.data('id')).replace('\\n',' ')) + '</li>';
-        }});
-        h += '</ul>';
-        document.getElementById('sidebarConnected').innerHTML = h;
-      }}
-
-      // Agent
-      if (t === 'agent') {{
-        var meta = '';
-        if (d.agentType) meta += 'Type: ' + d.agentType + '\\n';
-        if (d.source) meta += 'Source: ' + d.source + '\\n';
-        if (d.configPath) meta += 'Config: ' + d.configPath;
-        document.getElementById('sidebarMeta').textContent = meta;
-        var s = '<div class="sidebar-label">Statistics</div><ul class="sidebar-list">';
-        s += '<li>Servers: ' + (d.serverCount || 0) + '</li>';
-        s += '<li>Packages: ' + (d.packageCount || 0) + '</li>';
-        if (d.vulnCount) s += '<li style="color:#f87171">Vulnerabilities: ' + d.vulnCount + '</li>';
-        s += '</ul>';
-        document.getElementById('sidebarRemediation').innerHTML = s;
-      }}
-
-      // Server
-      if (t.indexOf('server_')===0) {{
-        if (d.command) document.getElementById('sidebarMeta').textContent = d.command;
-        var creds = []; try {{ creds = JSON.parse(d.credentials || '[]'); }} catch(e) {{}}
-        if (creds.length > 0) {{
-          var ch = '<div class="sidebar-label">Credentials (' + creds.length + ')</div><ul class="sidebar-list">';
-          creds.forEach(function(c) {{ ch += '<li>&#x1f511; <span class="sidebar-cred">' + escHtml(c) + '</span></li>'; }});
-          ch += '</ul>';
-          document.getElementById('sidebarCredentials').innerHTML = ch;
-        }}
-        var tools = []; try {{ tools = JSON.parse(d.toolNames || '[]'); }} catch(e) {{}}
-        if (tools.length > 0) {{
-          var th = '<div class="sidebar-label">MCP Tools (' + tools.length + ')</div><ul class="sidebar-list">';
-          tools.forEach(function(tl) {{ th += '<li>&#x1f527; ' + escHtml(tl) + '</li>'; }});
-          th += '</ul>';
-          document.getElementById('sidebarRemediation').innerHTML = th;
-        }}
-        var ph = '<div class="sidebar-label">Packages</div><ul class="sidebar-list">';
-        ph += '<li>Total: ' + (d.packageCount || 0) + '</li>';
-        if (d.vulnCount) ph += '<li style="color:#f87171">Vulnerable: ' + d.vulnCount + '</li>';
-        ph += '</ul>';
-        document.getElementById('sidebarCves').innerHTML = ph;
-      }}
-
-      // Package
-      if (t === 'pkg_vuln') {{
-        document.getElementById('sidebarMeta').textContent = (d.ecosystem || '') + ' \\u00b7 ' + (d.version || '');
-        var vids = []; try {{ vids = JSON.parse(d.vulnIds || '[]'); }} catch(e) {{}}
-        if (vids.length > 0) {{
-          var vh = '<div class="sidebar-label">CVEs (' + vids.length + ')</div><ul class="sidebar-list">';
-          vids.forEach(function(vid) {{
-            vh += '<li><a class="sidebar-link" href="https://osv.dev/vulnerability/' + urlPart(vid) + '" target="_blank" rel="noopener noreferrer">' + escHtml(vid) + ' &#x2197;</a></li>';
-          }});
-          vh += '</ul>';
-          document.getElementById('sidebarCves').innerHTML = vh;
-        }}
-      }}
-
-      // CVE
-      if (isCveNode) {{
-        var sev = d.severity || t.replace('cve_', '');
-        var mp = [];
-        if (sev) mp.push('Severity: ' + sev.toUpperCase());
-        if (d.cvssScore) mp.push('CVSS: ' + d.cvssScore);
-        document.getElementById('sidebarMeta').textContent = mp.join(' \\u00b7 ');
-        if (d.summary) {{
-          document.getElementById('sidebarCredentials').innerHTML = '<div class="sidebar-label">Summary</div><p style="font-size:.8rem;color:#cbd5e1;margin:0">' + escHtml(d.summary) + '</p>';
-        }}
-        var rh = '<div class="sidebar-label">Remediation</div><ul class="sidebar-list">';
-        if (d.fixVersion) {{
-          rh += '<li style="color:#4ade80">&#x2705; Fix: upgrade to <code>' + escHtml(d.fixVersion) + '</code></li>';
-        }} else {{
-          rh += '<li style="color:#f59e0b">&#x26a0; No fix available</li>';
-        }}
-        var lbl = d.label || '';
-        rh += '<li><a class="sidebar-link" href="https://osv.dev/vulnerability/' + urlPart(lbl) + '" target="_blank" rel="noopener noreferrer">View on OSV &#x2197;</a></li>';
-        rh += '<li><a class="sidebar-link" href="https://nvd.nist.gov/vuln/detail/' + urlPart(lbl) + '" target="_blank" rel="noopener noreferrer">View on NVD &#x2197;</a></li>';
-        rh += '</ul>';
-        document.getElementById('sidebarRemediation').innerHTML = rh;
-      }}
-
-      sidebar.classList.add('open');
-      sidebar.style.display = 'block';
-    }}
-
-    function closeSidebar() {{
-      sidebar.classList.remove('open');
-      setTimeout(function() {{ sidebar.style.display = 'none'; }}, 250);
-    }}
-
-    sidebarCloseBtn.addEventListener('click', closeSidebar);
-
-    cy.on('tap', 'node', function(e) {{
-      cy.elements().removeClass('faded highlighted');
-      var hood = e.target.closedNeighborhood();
-      cy.elements().not(hood).addClass('faded');
-      e.target.addClass('highlighted');
-      showSidebar(e.target);
-    }});
-    cy.on('tap', function(e) {{
-      if (e.target === cy) {{
-        cy.elements().removeClass('faded highlighted');
-        closeSidebar();
-      }}
-    }});
-
-    // Graph controls
-    document.getElementById('zoomIn').addEventListener('click', function() {{
-      cy.zoom({{ level: cy.zoom() * 1.3, renderedPosition: {{ x: cy.width() / 2, y: cy.height() / 2 }} }});
-    }});
-    document.getElementById('zoomOut').addEventListener('click', function() {{
-      cy.zoom({{ level: cy.zoom() / 1.3, renderedPosition: {{ x: cy.width() / 2, y: cy.height() / 2 }} }});
-    }});
-    document.getElementById('fitBtn').addEventListener('click', function() {{
-      cy.fit(cy.elements(), 40);
-    }});
-    document.getElementById('fullscreenBtn').addEventListener('click', function() {{
-      var gc = document.querySelector('.graph-container');
-      if (!document.fullscreenElement) {{
-        gc.requestFullscreen().then(function() {{
-          setTimeout(function() {{ cy.resize(); cy.fit(cy.elements(), 50); }}, 100);
-        }}).catch(function() {{}});
-      }} else {{
-        document.exitFullscreen();
-      }}
-    }});
-    document.addEventListener('fullscreenchange', function() {{
-      if (!document.fullscreenElement) {{
-        setTimeout(function() {{ cy.resize(); cy.fit(cy.elements(), 40); }}, 100);
-      }}
-    }});
-    // Graph severity filter
-    document.querySelectorAll('.graph-sev-filter').forEach(function(cb) {{
-      cb.addEventListener('change', function() {{
-        var checked = Array.from(document.querySelectorAll('.graph-sev-filter:checked')).map(function(c) {{ return c.value; }});
-        cy.nodes().forEach(function(n) {{
-          var t = n.data('type') || '';
-          if (t === 'cve' || t.startsWith('cve_')) {{
-            var sev = n.data('severity') || t.replace('cve_', '');
-            if (checked.indexOf(sev) === -1) {{
-              n.style('display', 'none');
-              n.connectedEdges().style('display', 'none');
-            }} else {{
-              n.style('display', 'element');
-              n.connectedEdges().style('display', 'element');
-            }}
-          }}
-        }});
-      }});
-    }});
-
-    // Graph search
-    var graphSearchInput = document.getElementById('graphSearch');
-    if (graphSearchInput) {{
-      graphSearchInput.addEventListener('input', function() {{
-        var q = this.value.toLowerCase();
-        if (!q) {{
-          cy.elements().removeClass('faded highlighted');
-          return;
-        }}
-        cy.elements().removeClass('faded highlighted');
-        var matched = cy.nodes().filter(function(n) {{
-          return (n.data('label') || '').toLowerCase().indexOf(q) >= 0;
-        }});
-        if (matched.length > 0) {{
-          var hood = matched.closedNeighborhood();
-          cy.elements().not(hood).addClass('faded');
-          matched.addClass('highlighted');
-        }} else {{
-          cy.elements().addClass('faded');
-        }}
-      }});
-    }}
-
-    // Context menu (right-click on nodes)
-    var ctxMenu = document.createElement('div');
-    ctxMenu.className = 'cy-ctx-menu';
-    document.body.appendChild(ctxMenu);
-
-    function hideCtxMenu() {{ ctxMenu.classList.remove('show'); }}
-    document.addEventListener('click', hideCtxMenu);
-    document.addEventListener('scroll', hideCtxMenu);
-
-    cy.on('cxttap', 'node', function(e) {{
-      e.originalEvent.preventDefault();
-      var node = e.target;
-      var d = node.data();
-      var t = d.type || '';
-      var items = [];
-
-      // Focus neighborhood
-      items.push({{icon:'&#x1f50d;',label:'Focus neighborhood',action:function(){{
-        cy.elements().removeClass('faded highlighted');
-        var hood = node.closedNeighborhood();
-        cy.elements().not(hood).addClass('faded');
-        node.addClass('highlighted');
-        showSidebar(node);
-      }}}});
-      // Fit to node
-      items.push({{icon:'&#x1f4cd;',label:'Zoom to node',action:function(){{
-        cy.animate({{ fit: {{ eles: node.closedNeighborhood(), padding: 80 }}, duration: 400 }});
-      }}}});
-      // Highlight path to root
-      items.push({{icon:'&#x2b06;',label:'Trace to root',action:function(){{
-        cy.elements().removeClass('faded highlighted');
-        var path = node.predecessors().union(node);
-        cy.elements().not(path).addClass('faded');
-        path.nodes().addClass('highlighted');
-      }}}});
-      // Highlight downstream
-      items.push({{icon:'&#x2b07;',label:'Show downstream impact',action:function(){{
-        cy.elements().removeClass('faded highlighted');
-        var downstream = node.successors().union(node);
-        cy.elements().not(downstream).addClass('faded');
-        downstream.nodes().addClass('highlighted');
-      }}}});
-
-      // Vulnerability node: open in OSV
-      if (t === 'cve' || t.indexOf('cve_')===0) {{
-        var vid = d.label || '';
-        items.push({{sep:true}});
-        items.push({{icon:'&#x1f517;',label:'Open in OSV',action:function(){{
-          window.open('https://osv.dev/vulnerability/'+vid, '_blank');
-        }}}});
-        items.push({{icon:'&#x1f517;',label:'Open in NVD',action:function(){{
-          window.open('https://nvd.nist.gov/vuln/detail/'+vid, '_blank');
-        }}}});
-      }}
-
-      // Build menu HTML
-      ctxMenu.innerHTML = '';
-      items.forEach(function(item) {{
-        if (item.sep) {{
-          var sep = document.createElement('div');
-          sep.className = 'cy-ctx-sep';
-          ctxMenu.appendChild(sep);
-        }} else {{
-          var el = document.createElement('div');
-          el.className = 'cy-ctx-item';
-          el.innerHTML = '<span>'+item.icon+'</span> '+item.label;
-          el.addEventListener('click', function(ev) {{
-            ev.stopPropagation();
-            hideCtxMenu();
-            item.action();
-          }});
-          ctxMenu.appendChild(el);
-        }}
-      }});
-
-      var cx = e.originalEvent.clientX, cy2 = e.originalEvent.clientY;
-      ctxMenu.style.left = cx + 'px';
-      ctxMenu.style.top = cy2 + 'px';
-      ctxMenu.classList.add('show');
-    }});
-
-    // Minimap — render a small overview of the full graph
-    var minimapEl = document.createElement('div');
-    minimapEl.className = 'cy-minimap';
-    cyContainer.parentNode.appendChild(minimapEl);
-    var mmCanvas = document.createElement('canvas');
-    mmCanvas.width = 180; mmCanvas.height = 130;
-    minimapEl.appendChild(mmCanvas);
-
-    function drawMinimap() {{
-      var ctx2d = mmCanvas.getContext('2d');
-      ctx2d.clearRect(0, 0, 180, 130);
-      var bb = cy.elements().boundingBox();
-      if (!bb || bb.w === 0) return;
-      var scaleX = 170 / bb.w, scaleY = 120 / bb.h;
-      var sc = Math.min(scaleX, scaleY);
-      var offX = (180 - bb.w * sc) / 2 - bb.x1 * sc;
-      var offY = (130 - bb.h * sc) / 2 - bb.y1 * sc;
-
-      // Draw edges
-      ctx2d.strokeStyle = '#334155'; ctx2d.lineWidth = 0.5;
-      cy.edges().forEach(function(edge) {{
-        var sp = edge.sourceEndpoint(), tp = edge.targetEndpoint();
-        ctx2d.beginPath();
-        ctx2d.moveTo(sp.x * sc + offX, sp.y * sc + offY);
-        ctx2d.lineTo(tp.x * sc + offX, tp.y * sc + offY);
-        ctx2d.stroke();
-      }});
-
-      // Draw nodes
-      cy.nodes().forEach(function(n) {{
-        var pos = n.position(); var t = n.data('type') || '';
-        var colors = {{'provider':'#818cf8','agent':'#3b82f6','server_clean':'#10b981','server_cred':'#f59e0b','server_vuln':'#ef4444','pkg_vuln':'#dc2626'}};
-        ctx2d.fillStyle = colors[t] || (t === 'cve' || t.indexOf('cve_')===0 ? '#f87171' : '#64748b');
-        var nx = pos.x * sc + offX, ny = pos.y * sc + offY;
-        ctx2d.beginPath();
-        if (t === 'cve' || t.indexOf('cve_')===0) {{
-          // Diamond
-          ctx2d.moveTo(nx, ny - 4); ctx2d.lineTo(nx + 5, ny); ctx2d.lineTo(nx, ny + 4); ctx2d.lineTo(nx - 5, ny);
-        }} else {{
-          ctx2d.arc(nx, ny, 3, 0, Math.PI * 2);
-        }}
-        ctx2d.fill();
-      }});
-
-      // Viewport rectangle
-      var ext = cy.extent();
-      ctx2d.strokeStyle = '#60a5fa'; ctx2d.lineWidth = 1.5;
-      ctx2d.strokeRect(ext.x1 * sc + offX, ext.y1 * sc + offY, ext.w * sc, ext.h * sc);
-    }}
-
-    cy.on('render viewport', drawMinimap);
-    setTimeout(drawMinimap, 500);
-
-    // Click minimap to pan
-    mmCanvas.addEventListener('click', function(e) {{
-      var rect = mmCanvas.getBoundingClientRect();
-      var mx = e.clientX - rect.left, my = e.clientY - rect.top;
-      var bb = cy.elements().boundingBox();
-      if (!bb || bb.w === 0) return;
-      var scaleX = 170 / bb.w, scaleY = 120 / bb.h;
-      var sc = Math.min(scaleX, scaleY);
-      var offX = (180 - bb.w * sc) / 2 - bb.x1 * sc;
-      var offY = (130 - bb.h * sc) / 2 - bb.y1 * sc;
-      var targetX = (mx - offX) / sc, targetY = (my - offY) / sc;
-      cy.animate({{ center: {{ x: targetX, y: targetY }}, duration: 300 }});
-    }});
-
-    // Node statistics overlay
-    var nodeStats = document.createElement('div');
-    nodeStats.style.cssText = 'position:absolute;bottom:12px;right:12px;background:rgba(15,23,42,.9);border:1px solid #334155;border-radius:8px;padding:8px 14px;font-size:.72rem;color:#64748b;z-index:10;backdrop-filter:blur(8px)';
-    var nodeCounts = {{}};
-    cy.nodes().forEach(function(n) {{
-      var t = n.data('type') || 'other';
-      if (t === 'cve' || t.indexOf('cve_')===0) t = 'cve';
-      else if (t.indexOf('server_')===0) t = 'server';
-      nodeCounts[t] = (nodeCounts[t] || 0) + 1;
-    }});
-    var statsHTML = [];
-    if (nodeCounts.agent) statsHTML.push('<span style="color:#3b82f6">' + nodeCounts.agent + ' agents</span>');
-    if (nodeCounts.server) statsHTML.push('<span style="color:#10b981">' + nodeCounts.server + ' servers</span>');
-    if (nodeCounts.pkg_vuln) statsHTML.push('<span style="color:#dc2626">' + nodeCounts.pkg_vuln + ' packages</span>');
-    if (nodeCounts.cve) statsHTML.push('<span style="color:#f87171">' + nodeCounts.cve + ' CVEs</span>');
-    nodeStats.innerHTML = statsHTML.join(' &middot; ');
-    cyContainer.parentNode.appendChild(nodeStats);
-
-  }} else if (cyContainer) {{
-    cyContainer.innerHTML = '<div style="display:flex;align-items:center;justify-content:center;height:100%;color:#4ade80;font-size:.9rem">&#x2705; No supply chain nodes to display</div>';
-  }}
-
-  // Findings-table pagination, filtering, and tabs live in the standalone
-  // scale-report script (below) so they keep working even if the CDN chart /
-  // graph libraries fail to load — the common case for an offline or emailed
-  // report. Nothing table/tab related runs here.
-
-  // Cytoscape: CVE Attack Flow graph
-  var cyAtkContainer = document.getElementById('cyAttack');
-  if (cyAtkContainer && ATTACK_FLOW.length > 0) {{
-    var cyAtk = cytoscape({{
-      container: cyAtkContainer,
-      elements: ATTACK_FLOW,
-      style: [
-        {{
-          selector: 'node[type="cve"], node[type^="cve_"]',
-          style: {{
-            'shape': 'diamond',
-            'width': 120,
-            'height': 34,
-            'label': 'data(label)',
-            'font-size': '9px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'color': '#fecaca',
-            'background-color': '#991b1b',
-            'border-color': '#f87171',
-            'border-width': 2.5,
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="critical"], node[type="cve_critical"]',
-          style: {{
-            'background-color': '#7f1d1d',
-            'border-color': '#ef4444',
-            'border-width': 3,
-            'width': 130,
-            'height': 38,
-            'underlay-color': '#ef4444',
-            'underlay-padding': '6px',
-            'underlay-opacity': 0.15,
-            'underlay-shape': 'ellipse',
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="high"], node[type="cve_high"]',
-          style: {{
-            'background-color': '#9a3412',
-            'border-color': '#fb923c',
-            'color': '#fed7aa',
-            'underlay-color': '#fb923c',
-            'underlay-padding': '4px',
-            'underlay-opacity': 0.1,
-            'underlay-shape': 'ellipse',
-          }},
-        }},
-        {{
-          selector: 'node[type="cve"][severity="medium"], node[type="cve"][severity="low"], node[type="cve"][severity="none"], node[type="cve_medium"], node[type="cve_low"], node[type="cve_none"]',
-          style: {{
-            'background-color': '#854d0e',
-            'border-color': '#fbbf24',
-            'border-width': 1.5,
-            'color': '#fef08a',
-            'width': 100,
-            'height': 28,
-          }},
-        }},
-        {{
-          selector: 'node[type="pkg_vuln"]',
-          style: {{
-            'background-color': '#7f1d1d',
-            'border-color': '#dc2626',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fca5a5',
-            'font-size': '9px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 130,
-            'height': 38,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '120px',
-          }},
-        }},
-        {{
-          selector: 'node[type="server"]',
-          style: {{
-            'background-color': '#1e293b',
-            'border-color': '#475569',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#cbd5e1',
-            'font-size': '10px',
-            'font-weight': '600',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 36,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '110px',
-          }},
-        }},
-        {{
-          selector: 'node[type="credential"]',
-          style: {{
-            'background-color': '#78350f',
-            'border-color': '#fbbf24',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#fde68a',
-            'font-size': '9px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 100,
-            'height': 32,
-            'shape': 'hexagon',
-          }},
-        }},
-        {{
-          selector: 'node[type="tool"]',
-          style: {{
-            'background-color': '#312e81',
-            'border-color': '#818cf8',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#c7d2fe',
-            'font-size': '9px',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 100,
-            'height': 30,
-            'shape': 'round-tag',
-            'text-wrap': 'wrap',
-            'text-max-width': '90px',
-          }},
-        }},
-        {{
-          selector: 'node[type="agent"]',
-          style: {{
-            'background-color': '#1e3a8a',
-            'border-color': '#3b82f6',
-            'border-width': 2,
-            'label': 'data(label)',
-            'color': '#bfdbfe',
-            'font-size': '11px',
-            'font-weight': '700',
-            'text-valign': 'center',
-            'text-halign': 'center',
-            'width': 120,
-            'height': 38,
-            'shape': 'round-rectangle',
-            'text-wrap': 'wrap',
-            'text-max-width': '105px',
-          }},
-        }},
-        {{
-          selector: 'edge',
-          style: {{
-            'width': 1.8,
-            'line-color': '#334155',
-            'target-arrow-color': '#475569',
-            'target-arrow-shape': 'triangle',
-            'curve-style': 'bezier',
-            'arrow-scale': 0.8,
-          }},
-        }},
-        {{
-          selector: 'edge[type="exploits"]',
-          style: {{
-            'line-color': '#dc2626',
-            'target-arrow-color': '#ef4444',
-            'width': 2.5,
-          }},
-        }},
-        {{
-          selector: 'edge[type="runs_on"]',
-          style: {{
-            'line-color': '#475569',
-            'target-arrow-color': '#64748b',
-          }},
-        }},
-        {{
-          selector: 'edge[type="exposes"]',
-          style: {{
-            'line-color': '#f59e0b',
-            'target-arrow-color': '#fbbf24',
-            'line-style': 'dashed',
-            'line-dash-pattern': [6, 3],
-            'width': 2,
-          }},
-        }},
-        {{
-          selector: 'edge[type="reaches"]',
-          style: {{
-            'line-color': '#818cf8',
-            'target-arrow-color': '#a5b4fc',
-            'line-style': 'dashed',
-            'line-dash-pattern': [4, 4],
-          }},
-        }},
-        {{
-          selector: 'edge[type="compromises"]',
-          style: {{
-            'line-color': '#ef4444',
-            'target-arrow-color': '#f87171',
-            'line-style': 'dashed',
-            'line-dash-pattern': [8, 4],
-            'width': 2.5,
-          }},
-        }},
-        {{
-          selector: '.highlighted',
-          style: {{
-            'border-width': 4,
-            'border-color': '#f1f5f9',
-            'z-index': 999,
-          }},
-        }},
-        {{
-          selector: '.faded',
-          style: {{ 'opacity': 0.08 }},
-        }},
-      ],
-      layout: {{
-        name: 'dagre',
-        rankDir: 'LR',
-        nodeSep: 40,
-        rankSep: 100,
-        edgeSep: 12,
-        padding: 30,
-        animate: false,
-        fit: true,
-      }},
-      minZoom: 0.15,
-      maxZoom: 4,
-      wheelSensitivity: 0.3,
-    }});
-    cyAtk.ready(function() {{ cyAtk.fit(cyAtk.elements(), 40); }});
-
-    // Attack flow tooltip
-    cyAtk.on('mouseover', 'node', function(e) {{
-      var t = e.target.data('tip');
-      if (t) {{ tip.textContent = t; tip.style.display = 'block'; }}
-    }});
-    cyAtk.on('mousemove', function(e) {{
-      if (tip.style.display === 'block') {{
-        tip.style.left = (e.originalEvent.clientX + 14) + 'px';
-        tip.style.top  = (e.originalEvent.clientY + 14) + 'px';
-      }}
-    }});
-    cyAtk.on('mouseout', 'node', function() {{ tip.style.display = 'none'; }});
-
-    // Attack flow click to highlight
-    cyAtk.on('tap', 'node', function(e) {{
-      cyAtk.elements().removeClass('faded highlighted');
-      var hood = e.target.closedNeighborhood();
-      cyAtk.elements().not(hood).addClass('faded');
-      e.target.addClass('highlighted');
-    }});
-    cyAtk.on('tap', function(e) {{
-      if (e.target === cyAtk) {{
-        cyAtk.elements().removeClass('faded highlighted');
-      }}
-    }});
-
-    // Attack flow controls
-    var afZoomIn = document.getElementById('afZoomIn');
-    var afZoomOut = document.getElementById('afZoomOut');
-    var afFitBtn = document.getElementById('afFitBtn');
-    if (afZoomIn) afZoomIn.addEventListener('click', function() {{
-      cyAtk.zoom({{ level: cyAtk.zoom() * 1.3, renderedPosition: {{ x: cyAtk.width() / 2, y: cyAtk.height() / 2 }} }});
-    }});
-    if (afZoomOut) afZoomOut.addEventListener('click', function() {{
-      cyAtk.zoom({{ level: cyAtk.zoom() / 1.3, renderedPosition: {{ x: cyAtk.width() / 2, y: cyAtk.height() / 2 }} }});
-    }});
-    if (afFitBtn) afFitBtn.addEventListener('click', function() {{
-      cyAtk.fit(cyAtk.elements(), 40);
-    }});
-
-    // Animated dash flow on exploit/compromises edges
-    var dashOffset = 0;
-    function animateAttackEdges() {{
-      dashOffset = (dashOffset + 0.5) % 24;
-      cyAtk.edges('[type="exploits"],[type="compromises"]').forEach(function(edge) {{
-        edge.style('line-dash-offset', -dashOffset);
-      }});
-      requestAnimationFrame(animateAttackEdges);
-    }}
-    // Only animate if attack edges exist
-    if (cyAtk.edges('[type="exploits"],[type="compromises"]').length > 0) {{
-      // Set dashed style for animation
-      cyAtk.edges('[type="exploits"]').style({{
-        'line-style': 'dashed',
-        'line-dash-pattern': [8, 4],
-      }});
-      cyAtk.edges('[type="compromises"]').style({{
-        'line-style': 'dashed',
-        'line-dash-pattern': [10, 5],
-      }});
-      animateAttackEdges();
-    }}
-
-    // Attack flow node count stats
-    var afStats = document.createElement('div');
-    afStats.style.cssText = 'position:absolute;bottom:12px;right:12px;background:rgba(15,23,42,.9);border:1px solid #334155;border-radius:8px;padding:8px 14px;font-size:.72rem;color:#64748b;z-index:10;backdrop-filter:blur(8px)';
-    var afCounts = {{}};
-    cyAtk.nodes().forEach(function(n) {{
-      var t = n.data('type') || 'other';
-      if (t === 'cve' || t.indexOf('cve_')===0) t = 'cve';
-      afCounts[t] = (afCounts[t] || 0) + 1;
-    }});
-    var afParts = [];
-    if (afCounts.cve) afParts.push('<span style="color:#f87171">' + afCounts.cve + ' CVEs</span>');
-    if (afCounts.pkg_vuln) afParts.push('<span style="color:#dc2626">' + afCounts.pkg_vuln + ' pkgs</span>');
-    if (afCounts.server) afParts.push('<span style="color:#64748b">' + afCounts.server + ' servers</span>');
-    if (afCounts.credential) afParts.push('<span style="color:#fbbf24">' + afCounts.credential + ' creds</span>');
-    if (afCounts.tool) afParts.push('<span style="color:#818cf8">' + afCounts.tool + ' tools</span>');
-    if (afCounts.agent) afParts.push('<span style="color:#3b82f6">' + afCounts.agent + ' agents</span>');
-    afStats.innerHTML = afParts.join(' &middot; ');
-    cyAtkContainer.parentNode.appendChild(afStats);
-  }}
-
-  // Table sorting
-  document.querySelectorAll('.data-table.sortable th').forEach(function(th) {{
-    th.addEventListener('click', function() {{
-      var table = th.closest('table');
-      var tbody = table.querySelector('tbody');
-      var rows = Array.from(tbody.querySelectorAll('tr'));
-      var col = parseInt(th.getAttribute('data-col'));
-      var arrow = th.querySelector('.sort-arrow');
-      var asc = !arrow.classList.contains('asc');
-
-      table.querySelectorAll('.sort-arrow').forEach(function(a) {{ a.className = 'sort-arrow'; }});
-      arrow.className = 'sort-arrow ' + (asc ? 'asc' : 'desc');
-
-      rows.sort(function(a, b) {{
-        var at = (a.children[col] || {{}}).textContent || '';
-        var bt = (b.children[col] || {{}}).textContent || '';
-        var an = parseFloat(at.replace(/[^\\d.-]/g, ''));
-        var bn = parseFloat(bt.replace(/[^\\d.-]/g, ''));
-        if (!isNaN(an) && !isNaN(bn)) return asc ? an - bn : bn - an;
-        return asc ? at.localeCompare(bt) : bt.localeCompare(at);
-      }});
-      rows.forEach(function(r) {{ tbody.appendChild(r); }});
-      // Re-slice the current page after re-ordering the DOM (paginator lives in
-      // the standalone scale-report script; guard in case it did not load).
-      if (window.PAGINATORS && window.PAGINATORS[table.id]) window.PAGINATORS[table.id].resort();
-    }});
-  }});
-
-  // Inventory search
-  var searchInput = document.getElementById('invSearch');
-  if (searchInput) {{
-    searchInput.addEventListener('input', function() {{
-      var q = this.value.toLowerCase();
-      document.querySelectorAll('.agent-card').forEach(function(card) {{
-        var text = card.textContent.toLowerCase();
-        card.style.display = text.includes(q) ? '' : 'none';
-      }});
-    }});
-  }}
-
-  // Package list toggle
-  window.togglePkgs = function(id, btn) {{
-    var el = document.getElementById(id);
-    if (!el) return;
-    var hidden = el.style.display === 'none';
-    el.style.display = hidden ? 'block' : 'none';
-    btn.innerHTML = hidden
-      ? 'Show fewer &#x25B2;'
-      : btn.dataset.orig || btn.innerHTML;
-    if (hidden && !btn.dataset.orig) btn.dataset.orig = btn.innerHTML;
-  }};
-
-  // Smooth scroll + close mobile sidebar. Tab reveal + robust anchor handling
-  // live in the standalone scale-report script so they survive CDN/graph load
-  // failures (offline / emailed reports).
-  document.querySelectorAll('a[href^="#"]').forEach(function(a) {{
-    a.addEventListener('click', function() {{
-      var sb = document.getElementById('mainSidebar');
-      if (sb) sb.classList.remove('mobile-open');
-    }});
-  }});
-
-  // Sidebar active section tracking via IntersectionObserver
-  var sidebarLinks = document.querySelectorAll('.sidebar-link');
-  var sections = document.querySelectorAll('section[id]');
-  if (sections.length > 0 && 'IntersectionObserver' in window) {{
-    var observer = new IntersectionObserver(function(entries) {{
-      entries.forEach(function(entry) {{
-        if (entry.isIntersecting) {{
-          sidebarLinks.forEach(function(link) {{
-            link.classList.remove('active');
-            if (link.getAttribute('href') === '#' + entry.target.id) {{
-              link.classList.add('active');
-            }}
-          }});
-        }}
-      }});
-    }}, {{ rootMargin: '-20% 0px -60% 0px', threshold: 0 }});
-    sections.forEach(function(sec) {{ observer.observe(sec); }});
-  }}
-}})();
-</script>
-
-</body>
-</html>"""
-    if offline_assets:
-        # Offline mode strips the interactive script layer, so keep the report a
-        # single scrollable page (no tab bar to click into a dead end).
-        return _apply_offline_assets_mode(html)
-    tab_counts = {
-        "findings": total_vulns + len(policy_findings),
-        "agents": len(report.agents),
-    }
-    return _apply_tabs(html, tab_counts)
-
-
-def export_html(
-    report: "AIBOMReport",
-    output_path: str,
-    blast_radii: list["BlastRadius"] | None = None,
-    *,
-    offline_assets: bool = False,
-) -> None:
-    """Write the HTML report to a file."""
-    Path(output_path).write_text(to_html(report, blast_radii or [], offline_assets=offline_assets), encoding="utf-8")
